### 2024-08-09 #7093 @editorialbot
[PRE REVIEW]: Metasyn: Transparent Generation of Synthetic Tabular Data with Privacy Guarantees
**Submitting author:** @vankesteren (<a href="http://orcid.org/0000-0003-1548-1663">Erik-Jan van Kesteren</a>)
**Repository:** https://github.com/sodascience/metasyn
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.2
