### 2024-08-09 4b72a5a
Better colours in readme gif
