### 2024-08-10 #7093 @vankesteren
In case you still need it, I looked up a few people on the JOSS reviewer list and these suggestions came up:
- [misken](https://github.com/misken)
- [prdm0](https://github.com/prdm0)
