### 2024-08-10 3e5a962
Add JOSS submission status badge to readme
