### 2024-08-12 a3efc60
first pass on paper shortening
2230 -> 1465 words
