+++ docs/paper/paper.bib
@@ -1,7 +1,9 @@
 @misc{bates2019ons,
   title={ONS methodology working paper series number 16—Synthetic data pilot},
   author={Bates, AG and Spakulov{\'a}, I and Dove, I and Mealor, A},
-  year={2019}
+  year={2019},
+  url={https://www.ons.gov.uk/methodology/methodologicalpublications/generalmethodology/onsworkingpaperseries/onsmethodologyworkingpaperseriesnumber16syntheticdatapilot},
+  urldate={2024-08-12}
 }
 
 @inproceedings{dwork2006differential,
@@ -10,14 +12,16 @@
   booktitle={International colloquium on automata, languages, and programming},
   pages={1--12},
   year={2006},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/11787006_1}
 }
 
-@book{dewolf2012statistical,
+@book{hundepool2012statistical,
   title={Statistical disclosure control},
-  author={de Wolf, Peter-Paul},
+  author={Hundepool, Anco and Domingo-Ferrer, Josep and Franconi, Luisa and Giessing, Sarah and Nordholt, Eric Schulte and Spicer, Keith and De Wolf, Peter-Paul},
   year={2012},
-  publisher={Wiley \& Sons, Chichester}
+  publisher={Wiley \& Sons, Chichester},
+  doi={10.1002/9781118348239}
 }
 
 @article{sweeney2002k,
@@ -28,22 +32,17 @@
   number={05},
   pages={557--570},
   year={2002},
-  publisher={World Scientific}
+  publisher={World Scientific},
+  doi={10.1142/S0218488502001648}
 }
 
 @misc{bond2015guidelines,
-  title={Guidelines for Output Checking. Eurostat},
+  title={Guidelines for the checking of output based on microdata research},
+  publisher={Eurostat},
   author={Bond, S and Brandt, M and de Wolf, PP},
-  year={2015}
-}
-
-@article{dwork2010differential,
-  title={Differential privacy for statistics: What we know and what we want to learn},
-  author={Dwork, Cynthia and Smith, Adam},
-  journal={Journal of Privacy and Confidentiality},
-  volume={1},
-  number={2},
-  year={2010}
+  year={2015},
+  url={https://web.archive.org/web/20160408145718/http://dwbproject.org/export/sites/default/about/public_deliveraples/dwb_d11-8_synthetic-data_cta-ecta_output-checking-guidelines_final-reports.zip},
+  urldate={2024-08-12}
 }
 
 @book{hastie2009elements,
@@ -51,16 +50,8 @@
   author={Hastie, Trevor and Tibshirani, Robert and Friedman, Jerome H and Friedman, Jerome H},
   volume={2},
   year={2009},
-  publisher={Springer}
-}
-
-@inproceedings{akaike1973information,
-  title={Information theory and an extension of the maximum likelihood principle},
-  author={Akaike, H},
-  booktitle={2nd International Symposium on Information Theory},
-  pages={267--281},
-  year={1973},
-  organization={Akad{\'e}miai Kiad{\'o} Location Budapest, Hungary}
+  publisher={Springer},
+  doi={10.1007/978-0-387-84858-7}
 }
 
 @article{neath2012bayesian,
@@ -71,8 +62,10 @@
   number={2},
   pages={199--203},
   year={2012},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/wics.199}
 }
+
 @software{vink2024polars,
   author       = {Ritchie Vink and
                   Stijn de Gooijer and
@@ -131,7 +124,8 @@
   journal={Journal of statistical software},
[…]
