### 2024-08-12 f36f1aa
Shorten to 1382 words and build paper
