+++ docs/paper/paper.md
@@ -45,7 +45,7 @@ These choices enable the software to generate synthetic data with __privacy and 
 
 At its core, `metasyn` has three main functions:
 
-1. __Estimation__: Automatically select distributions and fit them to a properly formatted tabular dataset, optionally with additional privacy guarantees.
+1. __Estimation__: Fit a generative model to a properly formatted tabular dataset, optionally with additional privacy guarantees.
 2. __(De)serialization__: Create an intermediate representation of the fitted model for auditing, editing, and exporting.
 3. __Generation__: Generate new synthetic datasets based on a fitted model.
 
@@ -74,25 +74,24 @@ Model estimation starts with an appropriately pre-processed data frame, meaning 
 └─────┴────────┴─────┴────────┴──────────┘
 ```
 
-For each data type, a set of candidate distributions is fitted (see Table \autoref{tbl:dist}), and then `metasyn` selects the one with the lowest BIC [@neath2012bayesian]. For distributions where BIC computation is impossible (e.g., for the string data type) a pseudo-BIC is created that trades off fit and complexity of the underlying models.
+For each data type, a set of candidate distributions is fitted (see \autoref{tbl:dist}), and then `metasyn` selects the one with the lowest BIC [@neath2012bayesian]. For distributions where BIC computation is impossible (e.g., for the string data type) a pseudo-BIC is created that trades off fit and complexity of the underlying models.
 
 Table: \label{tbl:dist} Candidate distributions associated with data types in the core `metasyn` package.
 
-| Variable type | Example                | Candidate distributions                                            |
-| :------------ | :--------------------- | :----------------------------------------------------------------- |
-| categorical   | yes/no, country        | Categorical (Multinoulli), Constant                                |
-| continuous    | 1.0, 2.1, ...          | Uniform, Normal, LogNormal, TruncatedNormal, Exponential, Constant |
-| discrete      | 1, 2, ...              | Poisson, Uniform, Normal, TruncatedNormal, Categorical, Constant   |
-| string        | A108, C122, some words | Regex, Categorical, Faker, FreeText, Constant                      |
-| date/time     | 2021-01-13, 01:40:12   | Uniform, Constant                                                  |
+| Data type   | Candidate distributions                                            |
+| :---------- | :----------------------------------------------------------------- |
+| Categorical | Categorical, Constant                                              |
+| Continuous  | Uniform, Normal, LogNormal, TruncatedNormal, Exponential, Constant |
+| Discrete    | Poisson, Uniform, Normal, TruncatedNormal, Categorical, Constant   |
+| String      | Regex, Categorical, Faker, FreeText, Constant                      |
+| Date/time   | Uniform, Constant                                                  |
 
[…]
