### 2024-08-16 7d87ae8
Further shorten paper
