### 2024-08-22 8810c0f
Add VarSpec to remove the warnings from the examples
