### 2024-08-23 c102905
One line for readthedocs and tutorials next steps (#310)
