+++ metasyn/demo/dataset.py
@@ -75,6 +75,7 @@ def demo_file(name: str = "titanic") -> Path:
         - titanic (Included in pandas, but post-processed to contain more columns)
         - spaceship (CC-BY from https://www.kaggle.com/competitions/spaceship-titanic)
         - fruit (very basic example data from Polars)
+        - survey (European Social Survey round 11 Netherlands Human Values Scale questionnaire; see https://doi.org/10.21338/ess11e01_0)
 
     Arguments
     ---------
@@ -92,9 +93,11 @@ def demo_file(name: str = "titanic") -> Path:
         return files(__package__) / "demo_spaceship.csv"
     if name == "fruit":
         return files(__package__) / "demo_fruit.csv"
+    if name == "survey":
+        return files(__package__) / "demo_survey.csv"
 
     raise ValueError(
-        f"No demonstration dataset with name '{name}'. Options: titanic, spaceship, fruit."
+        f"No demonstration dataset with name '{name}'. Options: titanic, spaceship, fruit, survey."
     )
 
 
@@ -105,6 +108,7 @@ def demo_dataframe(name: str = "titanic") -> pl.DataFrame:
         - titanic (Included in pandas, but post-processed to contain more columns)
         - spaceship (CC-BY from https://www.kaggle.com/competitions/spaceship-titanic)
         - fruit (very basic example data from Polars)
+        - survey (European Social Survey round 11 Netherlands Human Values Scale questionnaire; see https://doi.org/10.21338/ess11e01_0)
 
     Arguments
     ---------
@@ -126,9 +130,7 @@ def demo_dataframe(name: str = "titanic") -> pl.DataFrame:
             "Destination": pl.Categorical,
             "Transported": pl.Categorical,
         }
-        return pl.read_csv(
-            file_path, schema_overrides=data_types, try_parse_dates=True
-        )
+        return pl.read_csv(file_path, schema_overrides=data_types, try_parse_dates=True)
     if name == "titanic":
         # our edited titanic data
         data_types = {"Sex": pl.Categorical, "Embarked": pl.Categorical}
@@ -137,6 +139,8 @@ def demo_dataframe(name: str = "titanic") -> pl.DataFrame:
         # basic fruit data from polars example
         data_types = {"fruits": pl.Categorical, "cars": pl.Categorical}
         return pl.read_csv(file_path, schema_overrides=data_types)
+    if name == "survey":
+        return pl.read_csv(file_path)
 
     raise ValueError(
         f"No demonstration dataset with name '{name}'. Options: titanic, spaceship, fruit."
