+++ metasyn/demo/dataset.py
@@ -75,7 +75,7 @@ def demo_file(name: str = "titanic") -> Path:
         - titanic (Included in pandas, but post-processed to contain more columns)
         - spaceship (CC-BY from https://www.kaggle.com/competitions/spaceship-titanic)
         - fruit (very basic example data from Polars)
-        - survey (European Social Survey round 11 Netherlands Human Values Scale questionnaire; see https://doi.org/10.21338/ess11e01_0)
+        - survey (columns from ESS round 11 Human Values Scale questionnaire for the Netherlands, https://doi.org/10.21338/ess11e01_0)
 
     Arguments
     ---------
@@ -85,6 +85,10 @@ def demo_file(name: str = "titanic") -> Path:
     Returns
     -------
         Path to the dataset.
+
+    References
+    ----------
+    European Social Survey European Research Infrastructure (ESS ERIC). (2024). ESS11 integrated file, edition 1.0 [Data set]. Sikt - Norwegian Agency for Shared Services in Education and Research. https://doi.org/10.21338/ess11e01_0
 
     """
     if name == "titanic":
@@ -108,7 +112,7 @@ def demo_dataframe(name: str = "titanic") -> pl.DataFrame:
         - titanic (Included in pandas, but post-processed to contain more columns)
         - spaceship (CC-BY from https://www.kaggle.com/competitions/spaceship-titanic)
         - fruit (very basic example data from Polars)
-        - survey (European Social Survey round 11 Netherlands Human Values Scale questionnaire; see https://doi.org/10.21338/ess11e01_0)
+        - survey (columns from ESS round 11 Human Values Scale questionnaire for the Netherlands, https://doi.org/10.21338/ess11e01_0)
 
     Arguments
     ---------
@@ -118,6 +122,10 @@ def demo_dataframe(name: str = "titanic") -> pl.DataFrame:
     Returns
     -------
         Polars dataframe with correct column types
+
+    References
+    ----------
+    European Social Survey European Research Infrastructure (ESS ERIC). (2024). ESS11 integrated file, edition 1.0 [Data set]. Sikt - Norwegian Agency for Shared Services in Education and Research. https://doi.org/10.21338/ess11e01_0
 
     """
     file_path = demo_file(name=name)
