### 2024-08-30 5d99ac7
Format var.py to make ruff happy (hopefully!)
