+++ metasyn/var.py
@@ -12,7 +12,8 @@ from metasyn.privacy import BasePrivacy, BasicPrivacy
 from metasyn.provider import BaseDistributionProvider, DistributionProviderList
 from metasyn.varspec import DistributionSpec
 
-class MetaVar():
+
+class MetaVar:
     """Metadata variable describing a column in a MetaFrame.
 
     MetaVar is a structure that holds all metadata needed to generate a
@@ -49,14 +50,16 @@ class MetaVar():
         it will be assumed to have been created by the user.
     """
 
-    def __init__(self,  # pylint: disable=too-many-arguments
-                 name: str,
-                 var_type: str,
-                 distribution: BaseDistribution,
-                 dtype: str = "unknown",
-                 description: Optional[str] = None,
-                 prop_missing: float = 0.0,
-                 creation_method: Optional[dict] = None):
+    def __init__(
+        self,  # pylint: disable=too-many-arguments
+        name: str,
+        var_type: str,
+        distribution: BaseDistribution,
+        dtype: str = "unknown",
+        description: Optional[str] = None,
+        prop_missing: float = 0.0,
+        creation_method: Optional[dict] = None,
+    ):
         self.name = name
         self.var_type = var_type
         self.distribution = distribution
@@ -66,9 +69,11 @@ class MetaVar():
         self.creation_method = creation_method
         if creation_method is None:
             self.creation_method = {"created_by": "user"}
-        if self.prop_missing < -1e-8 or self.prop_missing > 1+1e-8:
-            raise ValueError(f"Cannot create variable '{self.name}' with proportion missing "
-                             "outside range [0, 1]")
+        if self.prop_missing < -1e-8 or self.prop_missing > 1 + 1e-8:
+            raise ValueError(
+                f"Cannot create variable '{self.name}' with proportion missing "
+                "outside range [0, 1]"
+            )
 
     @staticmethod
     def get_var_type(series: pl.Series) -> str:
@@ -111,8 +116,7 @@ class MetaVar():
         try:
             return convert_dict[polars_dtype]
         except KeyError as exc:
-            raise ValueError(
-                f"Unsupported polars type '{polars_dtype}'") from exc
+            raise ValueError(f"Unsupported polars type '{polars_dtype}'") from exc
 
     def to_dict(self) -> Dict[str, Any]:
         """Create a dictionary from the variable."""
@@ -145,21 +149,23 @@ class MetaVar():
 
         return (
             f'"{self.name}"\n'
-            f'{description}'
-            f'- Variable Type: {self.var_type}\n'
-            f'- Data Type: {self.dtype}\n'
-            f'- Proportion of Missing Values: {self.prop_missing:.4f}\n'
-            f'- Distribution:\n{distribution_formatted}\n'
+            f"{description}"
+            f"- Variable Type: {self.var_type}\n"
+            f"- Data Type: {self.dtype}\n"
+            f"- Proportion of Missing Values: {self.prop_missing:.4f}\n"
+            f"- Distribution:\n{distribution_formatted}\n"
         )
 
     @classmethod
-    def fit(cls,  # pylint: disable=too-many-arguments
-            series: pl.Series,
-            dist_spec: Optional[Union[dict, type, BaseDistribution, DistributionSpec]] = None,
-            provider_list: DistributionProviderList = DistributionProviderList("builtin"),
-            privacy: BasePrivacy = BasicPrivacy(),
-            prop_missing: Optional[float] = None,
-            description: Optional[str] = None) -> MetaVar:
+    def fit(
+        cls,  # pylint: disable=too-many-arguments
+        series: pl.Series,
+        dist_spec: Optional[Union[dict, type, BaseDistribution, DistributionSpec]] = None,
+        provider_list: DistributionProviderList = DistributionProviderList("builtin"),
[…]
