### 2024-08-30 82272bf
First try for simpler dev experience
- only use ruff for linting, which is now in a really good state to replace pydocstyle, flake8, pylint, etc
- separate github actions
- separate dependencies into test, check, extra, docs, and dev (which is everything)
- process: create changes, run ruff check, run mypy, run pytest, and you're good to commit!
