+++ .github/workflows/python-test.yml
@@ -1,7 +1,4 @@
-# This workflow will install Python dependencies, run tests and lint with a variety of Python versions
-# For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
-
-name: Lint, test, documentation, plugins
+name: Run tests on multiple platforms
 
 on:
   push:
@@ -12,17 +9,17 @@ on:
     - cron: "0 0 * * 0"
 
 jobs:
-  build:
+  test:
     strategy:
       fail-fast: false
       matrix:
         os: [ubuntu-latest]
-        python-version: [3.8, 3.9, "3.10", "3.11", "3.12"]
+        python-version: ["3.8", "3.9", "3.10", "3.11", "3.12"]
         include:
             - os: macos-latest
-              python-version: "3.11"
+              python-version: "3.12"
             - os: windows-latest
-              python-version: "3.11"
+              python-version: "3.12"
     runs-on: ${{ matrix.os }}
 
     steps:
@@ -34,35 +31,15 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        python -m pip install ".[test]"
-    - name: Lint with pylint
-      run: |
-        pylint metasyn
-    - name: Lint with Ruff
-      run: |
-        ruff check metasyn
-    - name: Check docstrings with pydocstyle
-      run: |
-        pydocstyle metasyn --convention=numpy --add-select=D417 --add-ignore="D102,D105"
-    - name: Check types with MyPy
-      run: |
-        mypy metasyn
-    - name: Check if documentation builds.
-      run: |
-        cd docs; make html SPHINXOPTS="--keep-going"
+        python -m pip install ".[test,extra]"
     - name: Test with pytest
-      run: |
-        pytest tests
+      run: pytest tests
     - name: Check notebook output
       if: ${{ matrix.os != 'macos-latest' }}
-      run: |
-        pytest --nbval-lax examples
-
+      run: pytest --nbval-lax examples
     - name: Test basic example
-      run: |
-        python examples/basic_example.py
-
-    - name: Test configuration example
+      run: python examples/basic_example.py
+    - name: Test CLI configuration example
       run: |
         pip install git+https://github.com/sodascience/metasyn-disclosure-control
         pip install .
