+++ .github/workflows/python-docs.yml
@@ -1,0 +1,25 @@
+name: Checking if documentation builds
+
+on:
+  push:
+    branches: [ main ]
+  pull_request:
+    branches: [ develop ]
+  schedule:
+    - cron: "0 0 * * 0"
+
+jobs:
+  docbuild:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v4
+    - name: Set up Python
+      uses: actions/setup-python@v4
+      with:
+        python-version: "3.x"
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        python -m pip install ".[docs]"
+    - name: Build docs
+      run: cd docs; make html SPHINXOPTS="--keep-going"

[…]
