### 2024-08-30 df92b0a
Update actions titles & setup-python version
There's a deprecation warning in setup-python@v4
