### 2024-08-30 12e567a
Add gif to readme (#304)
* Add gif to readme

fixes #250

* Better colours in readme gif

* Move gif up

* Little bump for new tests
