### 2024-08-30 sodascience/metasyn#317 @PetrKorab
Update README
Please, update the readme with a functional link (https://metasyn.readthedocs.io/en/latest/about/metasyn_in_detail.html). Either remove it or add the documentation. Thanks!
