### 2024-08-30 #7099 @PetrKorab
@vankesteren please, update readme [(#317)](https://github.com/sodascience/metasyn/issues/317). Thanks.
