+++ .github/workflows/python-test.yml
@@ -9,7 +9,29 @@ on:
     - cron: "0 0 * * 0"
 
 jobs:
+  full-check:
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: ["3.8", "3.9", "3.10", "3.11", "3.12"]
+    runs-on: "ubuntu-latest"
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          python -m pip install ".[check]"
+      - name: Lint with Ruff
+        run: ruff check
+      - name: Check types with MyPy
+        run: mypy metasyn  
+
   test:
+    needs: full-check # only run if lint and type check are ok
     strategy:
       fail-fast: false
       matrix:
