### 2024-09-02 17f543a
another round of simplification but now using jobs
