+++ .github/workflows/main-ci.yml
@@ -1,17 +1,16 @@
-name: Run tests on multiple platforms
+name: Continuous integration
 
 on:
   push:
-    branches: [ main ]
   pull_request:
-    branches: [ develop ]
   schedule:
     - cron: "0 0 * * 0"
 
 jobs:
-  full-check:
+  check:
+    name: Lint and check types
     strategy:
-      fail-fast: false
+      fail-fast: true
       matrix:
         python-version: ["3.8", "3.9", "3.10", "3.11", "3.12"]
     runs-on: "ubuntu-latest"
@@ -31,7 +30,8 @@ jobs:
         run: mypy metasyn  
 
   test:
-    needs: full-check # only run if lint and type check are ok
+    name: Run test suite
+    needs: check # only run if lint and type check are ok
     strategy:
       fail-fast: false
       matrix:
@@ -39,11 +39,10 @@ jobs:
         python-version: ["3.8", "3.9", "3.10", "3.11", "3.12"]
         include:
             - os: macos-latest
-              python-version: "3.12"
+              python-version: "3.x"
             - os: windows-latest
-              python-version: "3.12"
+              python-version: "3.x"
     runs-on: ${{ matrix.os }}
-
     steps:
     - uses: actions/checkout@v4
     - name: Set up Python ${{ matrix.python-version }}
@@ -66,3 +65,19 @@ jobs:
         pip install git+https://github.com/sodascience/metasyn-disclosure-control
         pip install .
         metasyn create-meta metasyn/demo/demo_titanic.csv --config examples/example_config.toml
+
+  build-docs:
+    name: Build documentation
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v4
+    - name: Set up Python
+      uses: actions/setup-python@v5
+      with:
+        python-version: "3.x"
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        python -m pip install ".[docs]"
+    - name: Build docs
+      run: cd docs; make html SPHINXOPTS="--keep-going"
