+++ .github/workflows/python-check.yml
@@ -1,20 +1,0 @@
-name: Linting and static type checking
-
-on: ["push", "pull_request"]
-jobs:
-  check:
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v4
-    - name: Set up Python
-      uses: actions/setup-python@v5
-      with:
-        python-version: "3.x"
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        python -m pip install ".[check]"
-    - name: Lint with Ruff
-      run: ruff check
-    - name: Check types with MyPy
-      run: mypy metasyn
