### 2024-09-02 d15651e
Add tests for data types
