### 2024-09-02 b935344
split dev and scheduled CI
