### 2024-09-02 2f7eac3
add pylint rules back in through ruff pyproject.toml
