+++ metasyn/distribution/base.py
@@ -158,7 +158,7 @@ class BaseDistribution(ABC):
         """Create a distribution from a dictionary."""
         return cls(**dist_dict["parameters"])
 
-    def information_criterion(self, values: Union[pl.Series, npt.NDArray]) -> float:  # pylint: disable=unused-argument
+    def information_criterion(self, values: Union[pl.Series, npt.NDArray]) -> float: # noqa: ARG002
         """Get the BIC value for a particular set of values.
 
         Parameters
@@ -350,7 +350,7 @@ class UniqueDistributionMixin(BaseDistribution):
             n_retry += 1
         raise ValueError(f"Failed to draw unique string after {n_retry} tries.")
 
-    def information_criterion(self, values):
+    def information_criterion(self, values): # noqa: ARG002
         return 9999999
 
 
