+++ pyproject.toml
@@ -92,7 +92,12 @@ ignore_missing_imports = true
 [tool.ruff]
 exclude = ["_version.py", "metasyn/distribution/legacy.*", "tests/*", "examples/*"]
 line-length = 100
-lint.select = ["E", "F", "W", "I", "N"]
+lint.select = ["E", "F", "W", "I", "N", "PLR", "ARG"]
+lint.ignore = ["PLR2004"]
 
 [tool.ruff.lint.pep8-naming]
 ignore-names = ["a", "b", "mu"]
+
+[tool.ruff.lint.pylint]
+max-args=10
+max-locals=35
