### 2024-09-03 10258e4
Rework demo datasets
