+++ metasyn/demo/dataset.py
@@ -1,6 +1,7 @@
 """Create and retrieve demo datasets."""
 
 # import random
+from abc import ABC, abstractclassmethod
 from pathlib import Path
 
 import polars as pl
@@ -10,62 +11,101 @@ try:
 except ImportError:
     from importlib.resources import files  # type: ignore
 
-# import numpy as np
-# import wget
-
-# from metasyn.distribution.datetime import (
-#     DateTimeUniformDistribution,
-#     DateUniformDistribution,
-#     TimeUniformDistribution,
-# )
+_AVAILABLE_DATASETS = {}
 
 
-# def create_titanic_demo(output_fp: Path) -> Path:
-#     """Create demo dataset for the titanic dataset.
+def register(*args):
+    def _wrap(cls):
+        _AVAILABLE_DATASETS[cls.name] = cls
+        return cls
+    return _wrap(*args)
 
-#     Arguments
-#     ---------
-#     output_fp:
-#         File to write the demonstration table to.
 
-#     Returns
-#     -------
-#         Output file location.
-#     """
-#     titanic_fp = Path("titanic.csv")
-#     if Path(output_fp).is_file():
-#         return output_fp
-#     if not titanic_fp.is_file():
-#         wget.download(
-#             "https://raw.githubusercontent.com/pandas-dev/pandas/main/doc/data/titanic.csv")
-#     dframe = pd.read_csv(titanic_fp)
-#     np.random.seed(1283742)
-#     random.seed(1928374)
+class BaseDataset(ABC):
 
-#     # Convert Age to a nullable integer.
-#     dframe["Age"] = dframe["Age"].round().astype("Int64")
+    @property
+    @abstractclassmethod
+    def name(cls):
+        pass
 
-#     # Add a date column.
-#     date_dist = DateUniformDistribution.default_distribution()
-#     dframe["Birthday"] = [date_dist.draw() if np.random.rand() < 0.9 else pd.NA
-#                           for _ in range(len(dframe))]
+    @classmethod
+    @property
+    def file_location(cls):
+        return files(__package__) / f"demo_{cls.name}.csv"
 
-#     # Add a time column.
+    @abstractclassmethod
+    def get_dataframe(cls):
+        return pl.read_csv(cls.file_location)
 
-#     time_dist = TimeUniformDistribution.default_distribution()
-#     dframe["Board time"] = [time_dist.draw() if np.random.rand() < 0.9 else pd.NA
-#                             for _ in range(len(dframe))]
 
-#     # Add a datetime column
-#     time_dist = DateTimeUniformDistribution.default_distribution()
-#     dframe["Married since"] = [time_dist.draw() if np.random.rand() < 0.9 else pd.NA
-#                                for _ in range(len(dframe))]
+@register
+class TitanicDataset(BaseDataset):
+    @classmethod
+    @property
+    def name(cls):
+        return "titanic"
 
-#     dframe["all_NA"] = [pd.NA for _ in range(len(dframe))]
-#     # Remove some columns for brevity and write to a file.
-#     dframe = dframe.drop(["SibSp", "Pclass", "Survived"], axis=1)
-#     dframe.to_csv(output_fp, index=False)
-#     return output_fp
+    @classmethod
+    def get_dataframe(cls):
+        file_path = cls.file_location
+        data_types = {"Sex": pl.Categorical, "Embarked": pl.Categorical}
+        return pl.read_csv(file_path, schema_overrides=data_types, try_parse_dates=True)
+
+
+@register
+class SpaceShipDataset(BaseDataset):
+    @classmethod
+    @property
+    def name(cls):
+        return "spaceship"
+
+    @classmethod
+    def get_dataframe(cls):
+        # the Kaggle spaceship data (CC-BY)
[…]
