### 2024-09-04 bb128f9
Bump version to 1.0.3
