+++ docs/source/conf.py
@@ -24,7 +24,7 @@ author = 'SODA Team'
 
 
 # The full version, including alpha/beta/rc tags
-release = '1.0.2'
+release = '1.0.3'
 
 
 # -- General configuration ---------------------------------------------------
