### 2024-09-04 bca2e01
Add all pylint rules, move dev status to beta
