+++ metasyn/distribution/categorical.py
@@ -14,6 +14,7 @@ from metasyn.distribution.base import BaseDistribution, metadist
 
 @metadist(implements="core.multinoulli", var_type=["categorical", "discrete", "string"])
 class MultinoulliDistribution(BaseDistribution):
+
     """Categorical distribution using labels and probabilities.
 
     This class represents a multinoulli (categorical) distribution.
@@ -34,6 +35,7 @@ class MultinoulliDistribution(BaseDistribution):
     --------
     >>> MultinoulliDistribution(labels=["a", "b", "b"], probs=[0.1, 0.3, 0.6])
     >>> MultinoulliDistribution(labels=[1, 3, 6], probs=[0.3, 0.4, 0.3])
+
     """
 
     def __init__(
