+++ metasyn/varspec.py
@@ -11,6 +11,7 @@ from metasyn.privacy import BasePrivacy, BasicPrivacy, get_privacy
 
 @dataclass
 class DistributionSpec():
+
     """Specification that determines which distribution is selected.
 
     It has the following attributes:
@@ -64,6 +65,7 @@ class DistributionSpec():
         ------
         TypeError
             If the input has the wrong type and cannot be parsed.
+
         """
         if isinstance(dist_spec, BaseDistribution):
             dist_dict = {key: value for key, value in dist_spec.to_dict().items()
@@ -90,6 +92,7 @@ class DistributionSpec():
         -------
             A flag that indicates whether a distribution can be generated from the values
             that are specified (not None).
+
         """
         return self.implements is not None and self.parameters is not None
 
@@ -104,6 +107,7 @@ class DistributionSpec():
         Returns
         -------
             Dictionary containing all the non-default settings for the creation method.
+
         """
         ret_dict: dict[str, Any] = {"created_by": "metasyn"}
         for var in ["implements", "unique", "parameters", "version"]:
@@ -117,6 +121,7 @@ class DistributionSpec():
 
 
 class VarSpec():  # pylint: disable=too-few-public-methods
+
     """Data class for storing the specifications for variables.
 
     Parameters
@@ -153,6 +158,7 @@ class VarSpec():  # pylint: disable=too-few-public-methods
         in the dataframe.
     var_type, optional:
         Manually set the variable type of the columns (used mainly for data_free columns).
+
     """
 
     def __init__(
@@ -196,5 +202,6 @@ class VarSpec():  # pylint: disable=too-few-public-methods
         Returns
         -------
             A VarSpec instance.
+
         """
         return cls(**var_dict)
