+++ metasyn/distribution/discrete.py
@@ -16,6 +16,7 @@ from metasyn.distribution.continuous import NormalDistribution, TruncatedNormalD
 
 @metadist(implements="core.uniform", var_type="discrete")
 class DiscreteUniformDistribution(ScipyDistribution):
+
     """Uniform discrete distribution.
 
     It differs from the floating point uniform distribution by
@@ -31,6 +32,7 @@ class DiscreteUniformDistribution(ScipyDistribution):
     Examples
     --------
     >>> DiscreteUniformDistribution(lower=3, upper=20)
+
     """
 
     dist_class = randint
@@ -60,6 +62,7 @@ class DiscreteUniformDistribution(ScipyDistribution):
 
 @metadist(implements="core.normal", var_type="discrete")
 class DiscreteNormalDistribution(NormalDistribution):
+
     """Normal discrete distribution.
 
     This class implements the normal/gaussian distribution and takes
@@ -76,6 +79,7 @@ class DiscreteNormalDistribution(NormalDistribution):
     Examples
     --------
     >>> DiscreteNormalDistribution(mean=2.4, sd=1.2)
+
     """
 
     def draw(self):
@@ -83,6 +87,7 @@ class DiscreteNormalDistribution(NormalDistribution):
 
 @metadist(implements="core.truncated_normal", var_type="discrete")
 class DiscreteTruncatedNormalDistribution(TruncatedNormalDistribution):
+
     """Truncated normal discrete distribution.
 
     Parameters
@@ -99,6 +104,7 @@ class DiscreteTruncatedNormalDistribution(TruncatedNormalDistribution):
     Examples
     --------
     >>> DiscreteTruncatedNormalDistribution(lower=1.2, upper=4.5, mean=2.3, sd=4.5)
+
     """
 
     def draw(self):
@@ -107,6 +113,7 @@ class DiscreteTruncatedNormalDistribution(TruncatedNormalDistribution):
 
 @metadist(implements="core.poisson", var_type="discrete")
 class PoissonDistribution(ScipyDistribution):
+
     """Poisson distribution.
 
     Parameters
@@ -117,6 +124,7 @@ class PoissonDistribution(ScipyDistribution):
     Examples
     --------
     >>> PoissonDistribution(rate=3.5)
+
     """
 
     dist_class = poisson
@@ -145,6 +153,7 @@ class PoissonDistribution(ScipyDistribution):
 
 @metadist(implements="core.unique_key", var_type="discrete", unique=True)
 class UniqueKeyDistribution(ScipyDistribution):
+
     """Unique key distribution for identifiers.
 
     Discrete distribution that ensures the uniqueness of the drawn values.
@@ -159,6 +168,7 @@ class UniqueKeyDistribution(ScipyDistribution):
     Examples
     --------
     >>> UniqueKeyDistribution(lower=0, consecutive=True)
+
     """
 
     def __init__(self, lower: int, consecutive: bool):
@@ -225,6 +235,7 @@ class UniqueKeyDistribution(ScipyDistribution):
 
 @metadist(implements="core.constant", var_type="discrete")
 class DiscreteConstantDistribution(BaseConstantDistribution):
+
     """Constant discrete distribution.
 
     This class implements the constant distribution, so that it draws always
@@ -238,6 +249,7 @@ class DiscreteConstantDistribution(BaseConstantDistribution):
     Examples
     --------
     >>> DiscreteConstantDistribution(213456)
+
     """
 
     @classmethod
