### 2024-09-04 11151d5
Update contributing guide for new dev flow
Also update sphinx requirement in pyproject.toml
