### 2024-09-04 bb8dad3
simplify doc build command
