### 2024-09-04 sodascience/metasyn#321 @PetrKorab
Improve description of the package, add examples, and use cases across disciplines (JOSS paper)
The package has detailed documentation, but the paper is quite difficult to read for a non-specialist audience, mainly in the second part. Please add example use cases of the package for researchers in various areas and leave part of the technicalities for the documentation.

It will help researchers across disciplines better understand the benefits of the package. Thanks!
