### 2024-09-04 #7099 @PetrKorab
@vankesteren Please be less technical in the second part of the paper. Instead, add examples and use cases across disciplines. Thanks!  ([#321](url)).
