### 2024-09-05 cc0e8ed
Remove blank lines at the end of docstrings
