+++ metasyn/provider.py
@@ -101,7 +101,6 @@ class BaseDistributionProvider(ABC):
         -------
         list[Type[BaseDistribution]]:
             List of distributions with that variable type.
-
         """
         dist_list = []
         if use_legacy:
@@ -177,7 +176,6 @@ class DistributionProviderList():
         The order in which distribution providers are included matters.
         If a provider implements the same distribution at the same privacy level,
         then only the first will be taken into account.
-
     """
 
     def __init__(
@@ -222,7 +220,6 @@ class DistributionProviderList():
             the BIC (Bayesian Information Criterion).
         privacy:
             Level of privacy that will be used in the fit.
-
         """
         if dist_spec.distribution is not None:
             return dist_spec.distribution
@@ -241,7 +238,6 @@ class DistributionProviderList():
         Returns
         -------
             A distribution according to the variable specifications.
-
         """
         dist_spec = var_spec.dist_spec
         unique = dist_spec.unique if dist_spec.unique else False
@@ -273,7 +269,6 @@ class DistributionProviderList():
         -------
         BaseDistribution:
             Distribution fitted to the series.
-
         """
         if len(series.drop_nulls()) == 0:
             return NADistribution()
@@ -328,7 +323,6 @@ class DistributionProviderList():
         -------
         tuple[Type[BaseDistribution], dict[str, Any]]:
             A distribution and the arguments to create an instance.
-
         """
         if NADistribution.matches_name(dist_name):
             return NADistribution
@@ -422,7 +416,6 @@ class DistributionProviderList():
         -------
         BaseDistribution:
             Fitted distribution.
-
         """
         unique = dist_spec.unique
         unique = unique if unique else False
@@ -464,7 +457,6 @@ class DistributionProviderList():
         -------
         dist_list:
             List of distributions that fit the given constraints.
-
         """
         dist_list = []
         for dist_provider in self.dist_packages:
@@ -488,7 +480,6 @@ class DistributionProviderList():
         -------
         BaseDistribution:
             Distribution representing the dictionary.
-
         """
         dist_name = var_dict["distribution"]["implements"]
         version = var_dict["distribution"].get("version", "1.0")
@@ -528,7 +519,6 @@ def get_distribution_provider(provider: Union[str, type[
     -------
     BaseDistributionProvider:
         The distribution provider that was found.
-
     """
     if isinstance(provider, BaseDistributionProvider):
         return provider
