### 2024-09-05 42ffed6
Simplify CI & development experience (#316)
* Improving / simplifying dev experience

- only use ruff for linting, which is now in a really good state to replace pydocstyle, flake8, pylint, etc
- separate github actions
- separate dependencies into test, check, extra, docs, and dev (which is everything)
- process: create changes, run ruff check, run mypy, run pytest, and you're good to commit!

* Update actions titles & setup-python version

There's a deprecation warning in setup-python@v4

* add lint and type check on PR matrix before test

* another round of simplification but now using jobs
[…]
