+++ metasyn/demo/dataset.py
@@ -76,6 +76,7 @@ def demo_file(name: str = "titanic") -> Path:
         - spaceship (CC-BY from https://www.kaggle.com/competitions/spaceship-titanic)
         - fruit (very basic example data from Polars)
         - survey (columns from ESS round 11 Human Values Scale questionnaire for the Netherlands)
+
     Arguments
     ---------
     name:
@@ -90,7 +91,6 @@ def demo_file(name: str = "titanic") -> Path:
     European Social Survey European Research Infrastructure (ESS ERIC). (2024). ESS11 integrated
     file, edition 1.0 [Data set]. Sikt - Norwegian Agency for Shared Services in Education and
     Research. https://doi.org/10.21338/ess11e01_0
-
     """
     if name == "titanic":
         return files(__package__) / "demo_titanic.csv"
@@ -129,7 +129,6 @@ def demo_dataframe(name: str = "titanic") -> pl.DataFrame:
     European Social Survey European Research Infrastructure (ESS ERIC). (2024). ESS11 integrated
     file, edition 1.0 [Data set]. Sikt - Norwegian Agency for Shared Services in Education and
     Research. https://doi.org/10.21338/ess11e01_0
-
     """
     file_path = demo_file(name=name)
     if name == "spaceship":
