### 2024-09-05 sodascience/metasyn#321 @misken
I concur. The documentation is quite detailed and no need to try to put too much of it into the paper. Specific usage examples from different domains might help readers envision their own use cases for metasyn.
