### 2024-09-06 sodascience/metasyn#324 @vankesteren
Hi @misken,

Thank you for the feedback!

Yes, that is an excellent question & it's on our radar; we intend to further explore this in future work around this package. The workaround you mentioned would be what we currently recommend, but we are definitely thinking about functional (manually specified) constraints as a feature directly within metasyn.

As an example of a potential generic solution, you can check my exploratory gist on this topic using permutations (esp. the images at the bottom). This gist is programmed in Julia but gives a nice indication that many functional relationships may be possible here: https://gist.github.com/vankesteren/6c141f7cabcd3eb47292d78cfca1804d
