### 2024-09-06 sodascience/metasyn#324 @vankesteren
To be more specific about the workaround, in your case I would
- load the data,
- compute the difference between start and end date,
- remove the end date column,
- then run metasyn
- export the metaframe
- generate synthetic data
- compute end date from the difference column
- remove the difference column
