### 2024-09-06 sodascience/metasyn#324 @misken
Yes, that makes sense as it retains the actual differences. Thanks.
