### 2024-09-09 fc0aba8
Add new tests for demo datasets
