+++ metasyn/demo/dataset.py
@@ -18,6 +18,7 @@ _AVAILABLE_DATASETS = {}
 
 
 def register(*args):
+    """Register a dataset so that it can be found by name."""
     def _wrap(cls):
         _AVAILABLE_DATASETS[cls.name] = cls
         return cls
@@ -25,6 +26,7 @@ def register(*args):
 
 
 class BaseDataset(ABC):
+    """Base class for demo datasets."""
 
     @property
     @abstractclassmethod
@@ -48,6 +50,7 @@ class BaseDataset(ABC):
 
 @register
 class TitanicDataset(BaseDataset):
+    """Included in pandas, but post-processed to contain more columns."""
     @classmethod
     @property
     def name(cls):
@@ -61,6 +64,7 @@ class TitanicDataset(BaseDataset):
 
 @register
 class SpaceShipDataset(BaseDataset):
+    """CC-BY from https://www.kaggle.com/competitions/spaceship-titanic"""
     @classmethod
     @property
     def name(cls):
@@ -80,6 +84,7 @@ class SpaceShipDataset(BaseDataset):
 
 @register
 class FruitDataset(BaseDataset):
+    """Very basic example data from Polars"""
     @classmethod
     @property
     def name(cls):
@@ -93,6 +98,7 @@ class FruitDataset(BaseDataset):
 
 @register
 class SurveyDataset(BaseDataset):
+    """Columns from ESS round 11 Human Values Scale questionnaire for the Netherlands"""
     @classmethod
     @property
     def name(cls):
@@ -106,6 +112,7 @@ class SurveyDataset(BaseDataset):
 
 @register
 class TestDataset(BaseDataset):
+    """Test dataset with all supported data types."""
     @classmethod
     @property
     def name(cls):
