### 2024-09-09 e69ac8e
Obtain dtype from polars for now
