### 2024-09-09 1bd017d
Redo API since class properties are deprecated
