### 2024-09-09 4cc74a2
Avoid polars bug with numpy 2.0
