### 2024-09-09 d68929e
Re-add numpy 2.0 support
