### 2024-09-09 93f7ab0
Improved demo datasets (#325)
Adds more tests, and fixes several typing issues. Now, all data types should be correctly stored and be reusable.
