+++ metasyn/demo/dataset.py
@@ -1,71 +1,195 @@
 """Create and retrieve demo datasets."""
 
 # import random
+import string
+from abc import ABC, abstractproperty
+from datetime import date, datetime, time, timedelta
 from pathlib import Path
 
+import numpy as np
 import polars as pl
+
+from metasyn.varspec import VarSpec
 
 try:
     from importlib_resources import files
 except ImportError:
     from importlib.resources import files  # type: ignore
 
-# import numpy as np
-# import wget
-
-# from metasyn.distribution.datetime import (
-#     DateTimeUniformDistribution,
-#     DateUniformDistribution,
-#     TimeUniformDistribution,
-# )
+_AVAILABLE_DATASETS = {}
 
 
-# def create_titanic_demo(output_fp: Path) -> Path:
-#     """Create demo dataset for the titanic dataset.
+def register(*args):
+    """Register a dataset so that it can be found by name."""
+    def _wrap(cls):
+        _AVAILABLE_DATASETS[cls().name] = cls()
+        return cls
+    return _wrap(*args)
 
-#     Arguments
-#     ---------
-#     output_fp:
-#         File to write the demonstration table to.
 
-#     Returns
-#     -------
-#         Output file location.
-#     """
-#     titanic_fp = Path("titanic.csv")
-#     if Path(output_fp).is_file():
-#         return output_fp
-#     if not titanic_fp.is_file():
-#         wget.download(
-#             "https://raw.githubusercontent.com/pandas-dev/pandas/main/doc/data/titanic.csv")
-#     dframe = pd.read_csv(titanic_fp)
-#     np.random.seed(1283742)
-#     random.seed(1928374)
+class BaseDataset(ABC):
+    """Base class for demo datasets."""
 
-#     # Convert Age to a nullable integer.
-#     dframe["Age"] = dframe["Age"].round().astype("Int64")
+    @abstractproperty
+    def name(self):
+        pass
 
-#     # Add a date column.
-#     date_dist = DateUniformDistribution.default_distribution()
-#     dframe["Birthday"] = [date_dist.draw() if np.random.rand() < 0.9 else pd.NA
-#                           for _ in range(len(dframe))]
+    @property
+    def file_location(self):
+        return files(__package__) / f"demo_{self.name}.csv"
 
-#     # Add a time column.
+    def get_dataframe(self):
+        return pl.read_csv(self.file_location, schema_overrides=self.schema,
+                           try_parse_dates=True)
 
-#     time_dist = TimeUniformDistribution.default_distribution()
-#     dframe["Board time"] = [time_dist.draw() if np.random.rand() < 0.9 else pd.NA
-#                             for _ in range(len(dframe))]
+    @abstractproperty
+    def schema(self):
+        pass
 
-#     # Add a datetime column
-#     time_dist = DateTimeUniformDistribution.default_distribution()
-#     dframe["Married since"] = [time_dist.draw() if np.random.rand() < 0.9 else pd.NA
-#                                for _ in range(len(dframe))]
+    @property
+    def var_specs(self):
+        return []
 
-#     dframe["all_NA"] = [pd.NA for _ in range(len(dframe))]
-#     # Remove some columns for brevity and write to a file.
-#     dframe = dframe.drop(["SibSp", "Pclass", "Survived"], axis=1)
-#     dframe.to_csv(output_fp, index=False)
-#     return output_fp
+
+@register
+class TitanicDataset(BaseDataset):
+    """Included in pandas, but post-processed to contain more columns."""
+
+    @property
+    def name(self):
+        return "titanic"
+
+    @property
+    def schema(self):
+        return {"Sex": pl.Categorical, "Embarked": pl.Categorical}
+
+    @property
[…]
