### 2024-09-09 sodascience/metasyn#321 @misken
Just to follow up with a bit more on this issue. In addition to some additional specific usage examples in the paper, it would be a good future task to base some of the documentation examples on more realistic datasets. I know that many people are familiar with the Titanic dataset and it does have fields that fall into the category of things that might be necessary to obscure via tools like `metasyn`. However, it might really help potential users to see usage examples based on your real experience with using `metasyn` to create synthetic data. The fruit based examples don't really scream out for the need for synthetic data. :)
