### 2024-09-17 b0d9621
Simplified configuration files
