+++ metasyn/provider.py
@@ -241,7 +241,6 @@ class DistributionProviderList():
         """
         dist_spec = var_spec.dist_spec
         unique = dist_spec.unique if dist_spec.unique else False
-        assert dist_spec.implements is not None and var_spec.var_type is not None
         dist_class = self.find_distribution(
             dist_spec.implements, var_spec.var_type,
             privacy=BasicPrivacy(), unique=unique)
@@ -298,7 +297,7 @@ class DistributionProviderList():
 
     def find_distribution(self,
                           dist_name: str,
-                          var_type: str,
+                          var_type: Optional[str],
                           privacy: Optional[BasePrivacy] = BasicPrivacy(),
                           unique: bool = False,
                           version: Optional[str] = None) -> type[BaseDistribution]:
