+++ metasyn/var.py
@@ -53,7 +53,7 @@ class MetaVar:
     def __init__( # noqa: PLR0913
         self,
         name: str,
-        var_type: str,
+        var_type: Optional[str],
         distribution: BaseDistribution,
         dtype: str = "unknown",
         description: Optional[str] = None,
@@ -61,6 +61,11 @@ class MetaVar:
         creation_method: Optional[dict] = None,
     ):
         self.name = name
+        if var_type is None:
+            if not isinstance(distribution.var_type, str):
+                raise ValueError("Failed to infer variable type for variable {name}"
+                                 " supply var_type or a different distribution.")
+            var_type = distribution.var_type
         self.var_type = var_type
         self.distribution = distribution
         self.dtype = dtype
