+++ metasyn/config.py
@@ -11,7 +11,7 @@ except ImportError:
 
 from metasyn.privacy import BasePrivacy, BasicPrivacy, get_privacy
 from metasyn.provider import DistributionProviderList
-from metasyn.varspec import VarSpec
+from metasyn.varspec import VarDefaults, VarSpec, DistributionSpec
 
 
 class MetaConfig():
@@ -44,32 +44,18 @@ class MetaConfig():
             self,
             var_specs: Union[list[dict], list[VarSpec]],
             dist_providers: Union[DistributionProviderList, list[str], str, None],
-            privacy: Optional[Union[BasePrivacy, dict]],
+            defaults: Optional[dict] = None,
             n_rows: Optional[int] = None):
         self.var_specs = [self._parse_var_spec(v) for v in var_specs]
         self.dist_providers = dist_providers  # type: ignore
-        self.privacy = privacy  # type: ignore
         self.n_rows = n_rows
+        self.defaults = VarDefaults(**defaults)
 
     @staticmethod
     def _parse_var_spec(var_spec):
         if isinstance(var_spec, VarSpec):
             return var_spec
         return VarSpec.from_dict(var_spec)
-
-    @property
-    def privacy(self) -> BasePrivacy:
-        """Return default privacy for generating the metaframe."""
-        return self._privacy
-
-    @privacy.setter
-    def privacy(self, privacy: Optional[Union[dict, BasePrivacy]]):
-        if privacy is None:
-            self._privacy: BasePrivacy = BasicPrivacy()
-        elif isinstance(privacy, dict):
-            self._privacy = get_privacy(**privacy)
-        else:
-            self._privacy = privacy
 
     @property
     def dist_providers(self) -> DistributionProviderList:
@@ -115,11 +101,11 @@ class MetaConfig():
         var_list = config_dict.pop("var", [])
         n_rows = config_dict.pop("n_rows", None)
         dist_providers = config_dict.pop("dist_providers", ["builtin"])
-        privacy = config_dict.pop("privacy", {"name": "none", "parameters": {}})
+        defaults = config_dict.pop("defaults", None)
         if len(config_dict) > 0:
             raise ValueError(f"Error parsing configuration file '{config_fp}'."
                              f" Unknown keys detected: '{list(config_dict)}'")
-        return cls(var_list, dist_providers, privacy, n_rows=n_rows)
+        return cls(var_list, dist_providers, defaults, n_rows=n_rows)
 
     def to_dict(self) -> dict:
         """Convert the configuration to a dictionary.
@@ -197,10 +183,19 @@ class VarSpecAccess():
         self.meta_config = meta_config
 
     def __getattribute__(self, attr):
-        if attr == "privacy":
-            if self.var_spec.privacy is None:
-                return self.meta_config.privacy
-            return self.var_spec.privacy
+        if attr in ["privacy", "data_free", "prop_missing", "privacy"]:
+            if getattr(self.var_spec, attr) is None:
+                return getattr(self.meta_config.defaults, attr)
+            return getattr(self.var_spec, attr)
+        if attr == "dist_spec":
+            if self.data_free and self.var_spec.dist_spec.implements is None:
+                if self.var_type not in self.meta_config.defaults.distribution:
+                    raise ValueError(f"Please supply distribution for variable with name {self.name}"
+                                      "that is datafree or set a default for the variable type"
+                                     f"{self.var_type}")
+                return DistributionSpec.parse(self.meta_config.defaults.distribution[self.var_type])
+            return self.var_spec.dist_spec
+
         if attr not in ("var_spec", "meta_config") and hasattr(self.var_spec, attr):
             return getattr(self.var_spec, attr)
         return super().__getattribute__(attr)
