### 2024-09-18 708bb31
Improve error messages and fix dtype bugs
