+++ metasyn/config.py
@@ -1,6 +1,7 @@
 """Module defining configuration classes for creating MetaFrames."""
 from __future__ import annotations
 
+import warnings
 from pathlib import Path
 from typing import Iterable, Optional, Union
 
@@ -9,9 +10,8 @@ try:
 except ImportError:
     import tomli as tomllib  # type: ignore  # noqa
 
-from metasyn.privacy import BasePrivacy, BasicPrivacy, get_privacy
 from metasyn.provider import DistributionProviderList
-from metasyn.varspec import VarDefaults, VarSpec, DistributionSpec
+from metasyn.varspec import DistributionSpec, VarDefaults, VarSpec
 
 
 class MetaConfig():
@@ -90,18 +90,27 @@ class MetaConfig():
             raise FileNotFoundError(f"It appears '{config_fp}' is not a valid filepath."
                                     f" Please provide a path to a .toml file to load a MetaConfig"
                                     f" from.") from fnf_error
-        except ValueError as value_error:
-            if Path(config_fp).suffix == ".toml":
+        except tomllib.TOMLDecodeError as value_error:
+            if Path(config_fp).suffix != ".toml":
                 raise ValueError(f"It appears '{Path(config_fp).name}' is a"
                                  f" '{Path(config_fp).suffix}' file."
                                  f" To load a MetaConfig, "
-                                 f"provide it as a .toml file.") from value_error
-            raise ValueError(f"An error occured while parsing the configuration file \n"
-                             f"('{Path(config_fp).name}').") from value_error
+                                 f"provide the configuration as a .toml file.") from value_error
+            raise value_error
         var_list = config_dict.pop("var", [])
         n_rows = config_dict.pop("n_rows", None)
         dist_providers = config_dict.pop("dist_providers", ["builtin"])
         defaults = config_dict.pop("defaults", None)
+        privacy = config_dict.pop("privacy", None)
+        config_version = config_dict.pop("config_version", "1.0")
+        if config_version not in ["1.0", "1.1"]:
+            warnings.warn(f"Trying to read configuration file with version {config_version}, "
+                          "this version of metasyn only supports 1.0 and 1.1.")
+        if privacy is not None:
+            if defaults is not None:
+                raise ValueError("Error parsing configuration file: cannot have both [privacy]"
+                                 " and [defaults] tables.")
+            defaults = {"privacy": privacy}
         if len(config_dict) > 0:
             raise ValueError(f"Error parsing configuration file '{config_fp}'."
                              f" Unknown keys detected: '{list(config_dict)}'")
@@ -190,9 +199,10 @@ class VarSpecAccess():
         if attr == "dist_spec":
             if self.data_free and self.var_spec.dist_spec.implements is None:
                 if self.var_type not in self.meta_config.defaults.distribution:
-                    raise ValueError(f"Please supply distribution for variable with name {self.name}"
-                                      "that is datafree or set a default for the variable type"
-                                     f"{self.var_type}")
+                    raise ValueError(
+                        f"Variable with name '{self.name}' is declared datafree and a distribution "
+                        f"cannot be inferred. You can set the distribution for '{self.name}' or set"
+                        f" a default distribution for variable type '{self.var_type}'.")
                 return DistributionSpec.parse(self.meta_config.defaults.distribution[self.var_type])
             return self.var_spec.dist_spec
 
