### 2024-09-18 sodascience/metasyn#326 @PetrKorab
add automated tests for main methods, add domain dataset(s)
Add automated tests covering the main Metasyn's methods, in line with [JOSS guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#tests). Also, please add tests for at least one domain dataset where you show the benefits of the programme. Make the data available to the reviewers, if possible. Thanks!
