### 2024-09-19 sodascience/metasyn#326 @qubixes
@PetrKorab Thanks for your comments. Could you elaborate in more detail which automated tests you are missing? We currently have >90% coverage and they are tested automatically with GitHub actions. Which main methods do you find untested?
