+++ metasyn/provider.py
@@ -241,6 +241,8 @@ class DistributionProviderList():
         """
         dist_spec = var_spec.dist_spec
         unique = dist_spec.unique if dist_spec.unique else False
+        if dist_spec.implements is None:
+            raise ValueError("Cannot create distribution without specifying the 'implements' key.")
         dist_class = self.find_distribution(
             dist_spec.implements, var_spec.var_type,
             privacy=BasicPrivacy(), unique=unique)
