+++ metasyn/privacy.py
@@ -51,6 +51,19 @@ class BasePrivacy(ABC):
         """
         return self.to_dict()["parameters"]
 
+    def comment(self, col_name):  # noqa
+        """Comment on the privacy features for the TOML GMF file.
+
+        Parameters
+        ----------
+        col_name
+            Name of the column.
+
+        Returns
+        -------
+            A comment on the privacy features.
+        """
+        return ""
 
 class BasicPrivacy(BasePrivacy):
     """Class representing no privacy level.
