+++ .github/workflows/core-ci.yml
@@ -21,7 +21,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          python -m pip install ".[check]"
+          python -m pip install ".[check]" tomlkit
       - name: Lint with Ruff
         run: ruff check
       - name: Check types with MyPy
