### 2024-09-19 sodascience/metasyn#326 @PetrKorab
@qubixes, Please add a link here to a folder with automated tests (JOSS submission). Thanks.
