### 2024-10-01 869ec96
Add tests and move gmf/config files
