+++ metasyn/testutils.py
@@ -7,10 +7,14 @@ and internal consistency of individual distributions and providers.
 
 from __future__ import annotations
 
+import json
+
 import jsonschema
+import numpy as np
 import polars as pl
 from jsonschema.exceptions import SchemaError
 
+from metasyn.distribution import MultinoulliDistribution, NADistribution
 from metasyn.distribution.base import BaseDistribution
 from metasyn.metaframe import _jsonify
 from metasyn.privacy import BasePrivacy
@@ -19,6 +23,7 @@ from metasyn.provider import (
     DistributionProviderList,
     get_distribution_provider,
 )
+from metasyn.var import MetaVar
 
 
 def check_distribution_provider(provider_name: str):
@@ -81,3 +86,114 @@ def check_distribution(distribution: type[BaseDistribution], privacy: BasePrivac
     assert isinstance(new_dist, distribution)
     assert set(list(new_dist.to_dict())) >= set(
         ("implements", "provenance", "class_name", "parameters"))
+
+
+
+HEADER_TEMPLATE = """
+# Metasyn report for {file_name}
+
+## General
+
+- Number of rows: {n_rows}
+- Number of columns: {n_columns}
+- Generated by {program_name} version {program_version} at {generation_time}
+
+## Variables / columns
+
+"""
+
+
+VAR_TEMPLATE = """
+### {var_name}
+
+- Distribution {class_name} with parameters:
+{parameters}
+- Variable type {var_type}
+- Based on {n_based_on} values{disclosure}
+- Percentage missing: {missing_perc} %
+- Generated examples: {example_list}
+
+
+"""
+
+def create_md_report(file_name, out_md_file):
+    with open(file_name, "r") as handle:
+        gmf_dict = json.load(handle)
+
+    header = HEADER_TEMPLATE.format(
+        file_name=file_name,
+        n_rows=gmf_dict["n_rows"],
+        n_columns=gmf_dict["n_columns"],
+        program_name=gmf_dict["provenance"]["created by"]["name"],
+        program_version=gmf_dict["provenance"]["created by"]["version"],
+        generation_time=gmf_dict["provenance"]["creation time"],
+    )
+    variables = ""
+
+    for var_dict in gmf_dict["vars"]:
+        var = MetaVar.from_dict(var_dict)
+        if isinstance(var.distribution, MultinoulliDistribution):
+            parameters = [f"\t- {label}: {round(prob*gmf_dict['n_rows']*(1-var.prop_missing))}\n"
+                          for label, prob in zip(var.distribution.labels, var.distribution.probs)]
+        elif isinstance(var.distribution, NADistribution):
+            variables += (f"### {var.name}\n- Distribution NADistribution\n- Only missing values"
+                          "\n- Examples: NA, NA, NA, ...\n")
+            continue
+        else:
+            parameters = [f"\t - {name}: {value}\n"
+                             for name, value in var.distribution._param_dict().items()]
+        parameter_str = "".join(parameters)
+        if var.prop_missing > 0:
+            examples = np.random.permutation([str(var.distribution.draw()) for _ in range(3)] +
+                                             ["NA", "NA"])
+        else:
+            examples = [str(x) for x in var.draw_series(5)]
+
+        if "privacy" in var_dict["creation_method"]:
+            partition_size = var_dict["creation_method"]["privacy"]["parameters"]["partition_size"]
+            disclosure = f" using micro aggregation with a partition size of {partition_size}"
+        else:
+            disclosure = ""
+        variables += VAR_TEMPLATE.format(
+            var_name = var.name,
+            var_type=var.var_type,
+            class_name=var.distribution.__class__.__name__,
[…]
