+++ examples/config_files/example_datafree.toml
@@ -1,0 +1,53 @@
+# Example toml file as input for metasyn
+
+config_version = "1.1"
+dist_providers = ["builtin"]
+n_rows = 100
+
+[defaults]
+
+data_free = true
+prop_missing = 0.1
+
+[defaults.distribution]
+discrete = {implements = "core.uniform", parameters = {lower = 1, upper = 30}}
+continuous = {implements = "core.normal", parameters = {mean = 0, stdev = 1}}
+string = {implements = "core.faker", parameters = {faker_type = "name", locale = "en_US"}}
+
+[[var]]
+name = "PassengerId"
+var_type = "discrete"
+distribution = {implements = "core.unique_key", unique = true, parameters = {"consecutive" = true, lower=1}}  # Notice booleans are lower case in .toml files.
+prop_missing = 0.0
+
+[[var]]
+name = "Name"
+description = "Name of the unfortunate passenger of the titanic."
+distribution = {implements = "core.faker", parameters = {faker_type = "name", locale = "en_US"}}
+
+[[var]]
+name = "Fare"
+distribution = {implements = "core.exponential", parameters = {rate = 0.4}}
+
+[[var]]
+name = "Age"
+var_type = "discrete"
+distribution = {implements = "core.uniform", parameters = {lower = 20, upper = 40}}
+
+[[var]]
+name = "Cabin"
+distribution = {implements = "core.regex", parameters = {regex_data = "[A-F][0-9]{2,3}"}}
+
+[[var]]
+name = "Secret"
+var_type = "discrete"
+
+
+[[var]]
+name = "Anotherone"
+var_type = "string"
+
+[[var]]
+name = "cats"
+var_type = "categorical"
+distribution = {implements = "core.multinoulli", parameters= {labels = ["male", "female"], probs=[0.45, 0.55]}}

[…]
