+++ tests/test_toml.py
@@ -1,10 +1,14 @@
-from metasyn.testutils import create_input_toml, create_md_report
-from metasyn.metaframe import MetaFrame
-from metasyn.provider import BuiltinDistributionProvider
-from metasyn.demo import demo_dataframe
 from filecmp import cmp
 from pathlib import Path
+
+import pytest
+import tomllib
 from pytest import mark
+
+from metasyn.demo import demo_dataframe
+from metasyn.metaframe import MetaFrame
+from metasyn.provider import BuiltinDistributionProvider
+from metasyn.testutils import create_input_toml, create_md_report
 
 
 def test_datafree_create(tmpdir):
@@ -13,7 +17,7 @@ def test_datafree_create(tmpdir):
     create_input_toml(temp_toml)
     assert cmp(temp_toml, Path("examples", "config_files", "example_all.toml"))
     mf = MetaFrame.fit_dataframe(None, var_specs=Path(temp_toml))
-    
+
     assert isinstance(mf, MetaFrame)
     assert mf.n_columns == len(BuiltinDistributionProvider.distributions)
 
@@ -26,8 +30,8 @@ def test_datafree_create(tmpdir):
 )
 def test_toml_store_load(tmpdir, toml_input, data):
     mf = MetaFrame.fit_dataframe(data, toml_input)
-    mf.to_toml(tmpdir/"test.toml")
-    new_mf = mf.from_toml(tmpdir/"test.toml")
+    mf.store(tmpdir/"test.toml")
+    new_mf = MetaFrame.load(tmpdir/"test.toml")
     assert mf.n_columns == new_mf.n_columns
 
 @mark.parametrize(
@@ -39,7 +43,7 @@ def test_toml_store_load(tmpdir, toml_input, data):
 def test_md_output(tmpdir, gmf_file):
     create_md_report(gmf_file, tmpdir/"test.md")
     assert Path(tmpdir/"test.md").is_file()
-    assert isinstance(MetaFrame.from_json(gmf_file), MetaFrame)
+    assert isinstance(MetaFrame.load(gmf_file), MetaFrame)
 
 def test_toml_err():
     with pytest.raises(FileNotFoundError):
