+++ metasyn/metaframe.py
@@ -6,7 +6,9 @@ import json
 import pathlib
 from datetime import datetime
 from importlib.metadata import version
+from pathlib import Path
 from typing import Any, Dict, List, Optional, Sequence, Union, no_type_check
+from warnings import warn
 
 import numpy as np
 import polars as pl
@@ -233,9 +235,54 @@ class MetaFrame():
             for i_desc, new_desc in enumerate(new_descriptions):
                 self[i_desc].description = new_desc
 
-    def export(self, fp: Optional[Union[pathlib.Path, str]],
-               validate: bool = True) -> None:
+    def store(self, fp: Optional[Union[pathlib.Path, str]],
+              validate: bool = True) -> None:
         """Serialize and export the MetaFrame to a JSON file, following the GMF format.
+
+        Optionally, validate the exported JSON file against the JSON schema(s) included in the
+        package.
+
+        Parameters
+        ----------
+        fp:
+            File to write the metaframe to.
+        validate:
+            Validate the JSON file with a schema.
+        """
+        fp_path = Path(fp)
+        if fp_path.suffix == ".toml":
+            self.store_toml(fp, validate)
+        else:
+            self.store_json(fp, validate)
+
+    @classmethod
+    def load(cls, fp: Optional[Union[pathlib.Path, str]], validate: bool = True) -> MetaFrame:
+        """Read a MetaFrame from a JSON file.
+
+        Parameters
+        ----------
+        fp:
+            Path to read the data from.
+        validate:
+            Validate the JSON file with a schema.
+
+        Returns
+        -------
+        MetaFrame:
+            A restored MetaFrame from the file.
+        """
+        fp_path = Path(fp)
+        if fp_path.suffix == ".toml":
+            return cls.load_toml(fp, validate)
+        else:
+            return cls.load_json(fp, validate)
+
+
+    def store_json(self, fp: Union[pathlib.Path, str],
+                validate: bool = True) -> None:
+        """Serialize and export the MetaFrame to a JSON file, following the GMF format.
+
+        This method is a wrapper and simply calls the 'export' function.
 
         Optionally, validate the exported JSON file against the JSON schema(s) included in the
         package.
@@ -256,27 +303,9 @@ class MetaFrame():
             with open(fp, "w", encoding="utf-8") as f:
                 json.dump(self_dict, f, indent=4)
 
-    def to_json(self, fp: Union[pathlib.Path, str],
-                validate: bool = True) -> None:
-        """Serialize and export the MetaFrame to a JSON file, following the GMF format.
-
-        This method is a wrapper and simply calls the 'export' function.
-
-        Optionally, validate the exported JSON file against the JSON schema(s) included in the
-        package.
-
-        Parameters
-        ----------
-        fp:
-            File to write the metaframe to.
-        validate:
-            Validate the JSON file with a schema.
-        """
-        self.export(fp, validate)
-
     @classmethod
-    def from_json(cls, fp: Union[pathlib.Path, str],
-                  validate: bool = True) -> MetaFrame:
+    def load_json(cls, fp: Union[pathlib.Path, str],
+                         validate: bool = True) -> MetaFrame:
         """Read a MetaFrame from a JSON file.
 
         Parameters
@@ -301,9 +330,32 @@ class MetaFrame():
         meta_vars = [MetaVar.from_dict(d) for d in self_dict["vars"]]
         return cls(meta_vars, n_rows)
 
+    def to_json(self, fp: Union[pathlib.Path, str], validate: bool = True) -> MetaFrame:
+        """Export, deprecated method, use Metaframe.export_to_json instead."""
[…]
