+++ tests/test_toml.py
@@ -2,7 +2,11 @@ from filecmp import cmp
 from pathlib import Path
 
 import pytest
-import tomllib
+try:
+    import tomllib
+except ImportError:
+    import tomli as tomllib
+
 from pytest import mark
 
 from metasyn.demo import demo_dataframe
