+++ examples/basic_example.py
@@ -17,12 +17,12 @@ mf = MetaFrame.fit_dataframe(df, var_specs=specs)
 
 # write to json
 gmf_path = Path("examples", "gmf_files", "example_gmf_simple.json")
-mf.export(gmf_path)
+mf.save(gmf_path)
 
 # then, export json from secure environment
 
 # outside secure environment, load json
-mf_out = MetaFrame.from_json(gmf_path)
+mf_out = MetaFrame.load_json(gmf_path)
 
 # create a fake dataset
 df_syn = mf_out.synthesize(10)
