### 2024-10-04 ee06827
Change the representation of MetaVar/MetaFrame
