### 2024-10-09 3938303
Add new dataset and make regex more used (#330)
* Add new dataset and make regex more used

* Fix linting issue

* Automatically add new datasets to tests

---------

Co-authored-by: Erik-Jan van Kesteren <erikjanvankesteren@pm.me>
