### 2024-10-09 789ae63
Change the representation of MetaVar/MetaFrame (#331)
* Change the representation of MetaVar/MetaFrame

* Remove old code
