### 2024-10-10 993e60b
Update .github/workflows/core-ci.yml
Co-authored-by: Erik-Jan van Kesteren <e.vankesteren1@uu.nl>
