+++ .github/workflows/core-ci.yml
@@ -21,7 +21,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          python -m pip install ".[check]" tomlkit
+          python -m pip install ".[check,extra]"
       - name: Lint with Ruff
         run: ruff check metasyn
       - name: Check types with MyPy
