### 2024-10-10 d63518c
Update metasyn/__main__.py
Co-authored-by: Erik-Jan van Kesteren <e.vankesteren1@uu.nl>
