### 2024-10-10 cd0e8ac
Add export functionality to TOML in addition to JSON (#327)
* Add support for simplified configuration file.
* Rename methods:
   * MetaFrame.export -> MetaFrame.save
   * MetaFrame.to_json -> MetaFrame.save_json
   * MetaFrame.from_json -> MetaFrame.load_json
* Add support for TOML GMF files. They work almost exactly the same as the JSON GMF files. Some comments are automatically included to make the file more easily understandable.
