+++ metasyn/metaframe.py
@@ -6,14 +6,22 @@ import json
 import pathlib
 from datetime import datetime
 from importlib.metadata import version
-from typing import Any, Dict, List, Optional, Sequence, Union
+from pathlib import Path
+from typing import Any, Dict, List, Optional, Sequence, Union, no_type_check
+from warnings import warn
 
 import numpy as np
 import polars as pl
+
+try:
+    import tomllib
+except ImportError:
+    import tomli as tomllib  # type: ignore  # noqa
+
 from tqdm import tqdm
 
 from metasyn.config import MetaConfig
-from metasyn.privacy import BasePrivacy
+from metasyn.privacy import BasePrivacy, get_privacy
 from metasyn.validation import validate_gmf_dict
 from metasyn.var import MetaVar
 from metasyn.varspec import VarSpec
@@ -23,7 +31,7 @@ class MetaFrame():
     """Container for statistical metadata describing a dataset.
 
     This class is used to fit a MetaFrame to a Polars DataFrame, serialize and
-    export the MetaFrame to a file, read a MetaFrame from a file, and create
+    save the MetaFrame to a file, read a MetaFrame from a file, and create
     a synthetic Polars DataFrame.
 
     A MetaFrame represents a metadata frame, which is a structure that holds
@@ -104,13 +112,13 @@ class MetaFrame():
         elif isinstance(var_specs, MetaConfig):
             meta_config = var_specs
         elif var_specs is None:
-            meta_config = MetaConfig([], dist_providers, privacy)
+            meta_config = MetaConfig([], dist_providers, defaults = {"privacy": privacy})
         else:
-            meta_config = MetaConfig(var_specs, dist_providers, privacy)
+            meta_config = MetaConfig(var_specs, dist_providers, defaults = {"privacy": privacy})
         if dist_providers is not None:
             meta_config.dist_providers = dist_providers  # type: ignore
         if privacy is not None:
-            meta_config.privacy = privacy  # type: ignore
+            meta_config.defaults["privacy"] = privacy  # type: ignore
 
         if df is not None and not isinstance(df, pl.DataFrame):
             if isinstance(df, (str, pathlib.Path)):
@@ -227,11 +235,61 @@ class MetaFrame():
             for i_desc, new_desc in enumerate(new_descriptions):
                 self[i_desc].description = new_desc
 
-    def export(self, fp: Optional[Union[pathlib.Path, str]],
-               validate: bool = True) -> None:
-        """Serialize and export the MetaFrame to a JSON file, following the GMF format.
+    def save(self, fp: Optional[Union[pathlib.Path, str]], validate: bool = True) -> None:
+        """Serialize and save the MetaFrame to a JSON or TOML file, following the GMF format.
 
-        Optionally, validate the exported JSON file against the JSON schema(s) included in the
+        Optionally, validate the saved JSON file against the JSON schema(s) included in the
+        package. A TOML cannot be validated against a schema currently.
+
+        Parameters
+        ----------
+        fp:
+            File to write the metaframe to.
+        validate:
+            Validate the JSON file with a schema. If the file is a TOML file, then this will
+            be ignored.
+        """
+        if fp is None:
+            self.save_json(fp, validate)
+            return
+        fp_path = Path(fp)
+        if fp_path.suffix == ".toml":
+            self.save_toml(fp, validate)
+        else:
+            self.save_json(fp, validate)
+
+    @classmethod
+    def load(cls, fp: Union[pathlib.Path, str], validate: bool = True) -> MetaFrame:
+        """Read a MetaFrame from a JSON or TOML GMF file.
+
+        Optionally, validate the saved JSON file against the JSON schema(s) included in the
[…]
