### 2024-10-11 #7099 @vankesteren
Hey @crvernon thanks for asking;

Until now, we have responded to several issues that have come out of this review (e.g., https://github.com/sodascience/metasyn/issues/317, https://github.com/sodascience/metasyn/issues/324) and we are working on changes based on the other issues. For example, we have included a medical example dataset in the package (https://github.com/sodascience/metasyn/pull/330) as a partial response to https://github.com/sodascience/metasyn/issues/321 and https://github.com/sodascience/metasyn/issues/326.

I have gotten in a bit of a time-crunch with teaching, but I'm committed to update the paper itself over the next weeks to include more examples and fewer technical details, for which we can refer to the documentation. In the meantime, we are also slightly updating the documentation to have a better flow through these technical details.

In summary, I haven't responded much over the past weeks but we have been hard at work! Let me know if you require something additional.
