### 2024-10-14 #7099 @PetrKorab
@vankesteren The medical dataset is ok, but for the reviewer, it is complicated to evaluate Metasyn without **well-documented** GitHub actions/ other CLI. Please place everything into a single folder. Thanks!
