### 2024-10-16 db65390
Fix scheduled CI
