+++ .github/workflows/scheduled-ci.yml
@@ -27,7 +27,7 @@ jobs:
           python -m pip install --upgrade pip
           python -m pip install ".[check]"
       - name: Lint with Ruff
-        run: ruff check
+        run: ruff check metasyn
       - name: Check types with MyPy
         run: mypy metasyn  
 
