### 2024-10-16 19e49e8
Update scheduled-ci.yml
