### 2024-10-16 708563b
Update scheduled-ci.yml
