+++ .github/workflows/scheduled-ci.yml
@@ -25,7 +25,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          python -m pip install ".[check]"
+          python -m pip install ".[check,extra]"
       - name: Lint with Ruff
         run: ruff check metasyn
       - name: Check types with MyPy
