### 2024-10-18 6f69e97
Improve the warning for uniqueness
Also lower the frequency of this warning. Only if we are quite certain
about the uniqueness, a warning will be issued.
