### 2024-10-18 e049918
Update schedule for Python 3.8
