+++ .github/workflows/scheduled-ci.yml
@@ -12,7 +12,7 @@ jobs:
       fail-fast: true
       matrix:
         branch: ["main", "develop"]
-        python-version: ["3.8", "3.9", "3.10", "3.11", "3.12"]
+        python-version: ["3.9", "3.10", "3.11", "3.12", "3.13"]
     runs-on: "ubuntu-latest"
     steps:
       - uses: actions/checkout@v4
@@ -38,7 +38,7 @@ jobs:
       matrix:
         branch: ["main", "develop"]
         os: [ubuntu-latest]
-        python-version: ["3.8", "3.9", "3.10", "3.11", "3.12"]
+        python-version: ["3.9", "3.10", "3.11", "3.12", "3.13"]
         include:
             - os: macos-latest
               python-version: "3.x"
