### 2024-10-21 3f6677e
Fix os x/windows CI
