+++ .github/workflows/core-ci.yml
@@ -36,9 +36,9 @@ jobs:
         python-version: ["3.9", "3.10", "3.11", "3.12"]
         include:
             - os: macos-latest
-              python-version: "3.x"
+              python-version: "3.12"
             - os: windows-latest
-              python-version: "3.x"
+              python-version: "3.12"
     runs-on: ${{ matrix.os }}
     steps:
     - uses: actions/checkout@v4
