### 2024-10-21 e020900
Improve unique warning (#332)
Lower the frequency of this warning. Only if we are quite certain
about the uniqueness, a warning will be issued. We now have an exception for unique key columns: if they are consecutive and metasyn is quite certain in it being the best fit, they will be chosen.
