+++ docs/source/faq.rst
@@ -22,8 +22,8 @@ The process of generating synthetic data solely from the MetaFrame ensures that 
 
 
 
-**I encountered the warning: "Variable PassengerId seems unique, but not set to be unique." What should I do?**
------------------------------------------------------------------------------------------------------------------
+**I encountered the warning: "Metasyn detected that variable {x} is potentially unique." What should I do?**
+------------------------------------------------------------------------------------------------------------
 This warning occurs when ``metasyn`` detects a column, that seems to have unique values in the real dataset but isn't specified to be unique in the fitting of the :obj:`MetaFrame<metasyn.metaframe.MetaFrame>`. To address this, you can use the `spec` parameter to create a specification dictionary and indicate that the column should have unique values. Here's an example of how to do this (in this example ``PassengerId`` is the column with unique values):
 
 .. code-block:: python
@@ -39,6 +39,8 @@ This warning occurs when ``metasyn`` detects a column, that seems to have unique
    mf = MetaFrame.fit_dataframe(df, var_specs=var_specs)
 
 More information on how to use the optional parameters in the :meth:`metasyn.MetaFrame.fit_dataframe() <metasyn.metaframe.MetaFrame.fit_dataframe>` function can be found in :doc:`/usage/generating_metaframes` under :ref:`optionalparams`.
+
+You can also set the uniqueness of a variable in the :doc:`configuration file <usage/config_files>`
 
 **I found a bug/issue, where can I report it?**
 -----------------------------------------------
