+++ .github/workflows/scheduled-ci.yml
@@ -12,7 +12,7 @@ jobs:
       fail-fast: false
       matrix:
         branch: ["main", "develop"]
-        python-version: ["3.8", "3.9", "3.10", "3.11", "3.12"]
+        python-version: ["3.9", "3.10", "3.11", "3.12"]
     runs-on: "ubuntu-latest"
     steps:
       - uses: actions/checkout@v4
@@ -38,12 +38,12 @@ jobs:
       matrix:
         branch: ["main", "develop"]
         os: [ubuntu-latest]
-        python-version: ["3.8", "3.9", "3.10", "3.11", "3.12"]
+        python-version: ["3.9", "3.10", "3.11", "3.12"]
         include:
             - os: macos-latest
-              python-version: "3.x"
+              python-version: "3.12"
             - os: windows-latest
-              python-version: "3.x"
+              python-version: "3.12"
     runs-on: ${{ matrix.os }}
     steps:
     - uses: actions/checkout@v4
