+++ .github/workflows/core-ci.yml
@@ -10,7 +10,7 @@ jobs:
     strategy:
       fail-fast: true
       matrix:
-        python-version: ["3.8", "3.9", "3.10", "3.11", "3.12"]
+        python-version: ["3.9", "3.10", "3.11", "3.12"]
     runs-on: "ubuntu-latest"
     steps:
       - uses: actions/checkout@v4
@@ -33,12 +33,12 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest]
-        python-version: ["3.8", "3.9", "3.10", "3.11", "3.12"]
+        python-version: ["3.9", "3.10", "3.11", "3.12"]
         include:
             - os: macos-latest
-              python-version: "3.x"
+              python-version: "3.12"
             - os: windows-latest
-              python-version: "3.x"
+              python-version: "3.12"
     runs-on: ${{ matrix.os }}
     steps:
     - uses: actions/checkout@v4
