### 2024-10-28 a8d51d8
Add more comments and change NA distribution
