+++ tests/test_var.py
@@ -1,3 +1,7 @@
+"""Module to test the MetaVar object.
+
+Most of the test functions try to infer the correct distribution from a series.
+"""
 import datetime as dt
 import json
 import string
@@ -37,6 +41,7 @@ def _series_element_classname(series, all_nan):
 
 
 def check_var(series, var_type, temp_path, all_nan=False):
+    """Checker for all variables whether series are inferred correctly."""
     def check_similar(series_a, series_b):
         assert isinstance(series_a, (pd.Series, pl.Series))
         assert isinstance(series_b, (pd.Series, pl.Series))
@@ -101,11 +106,13 @@ def check_var(series, var_type, temp_path, all_nan=False):
     ]
 )
 def test_categorical(tmp_path, series):
+    """Test the inference of categorical variables."""
     new_series = check_var(series, "categorical", tmp_path)
     assert set(_series_drop_nans(series)) == set(np.unique(_series_drop_nans(new_series)))
 
 @mark.parametrize("dtype", ["int8", "int16", "int32", "int64", "int"])
 def test_integer(dtype, tmp_path):
+    """Test the inference of integer variables."""
     series = pd.Series([np.random.randint(0, 10) for _ in range(300)], dtype=dtype)
     new_series = check_var(series, "discrete", tmp_path)
     assert new_series.min() >= 0
@@ -114,6 +121,7 @@ def test_integer(dtype, tmp_path):
 
 @mark.parametrize("dtype", ["Int8", "Int16", "Int32", "Int64"])
 def test_nullable_integer(dtype, tmp_path):
+    """Test for pandas nullable integer data types (no applicable to polars)."""
     series = pd.Series([np.random.randint(0, 10) if np.random.rand() > 0.5 else None
                         for _ in range(500)], dtype=dtype)
     new_series = check_var(series, "discrete", tmp_path)
@@ -125,6 +133,7 @@ def test_nullable_integer(dtype, tmp_path):
               pl.UInt8, pl.UInt16, pl.UInt32, pl.UInt64]
 )
 def test_polars_discrete(dtype, tmp_path):
+    """Test for all different variables of polars discrete types."""
     series = pl.Series("ser", [np.random.randint(0, 10) for _ in range(100)], dtype=dtype)
     new_series = check_var(series, "discrete", tmp_path)
     assert new_series.min() >= 0
@@ -134,6 +143,7 @@ def test_polars_discrete(dtype, tmp_path):
     "dtype", [pl.Float32, pl.Float64]
 )
 def test_polars_continuous(dtype, tmp_path):
+    """Test for all polars continuous data types."""
     series = pl.Series("ser", np.random.rand(100), dtype=dtype)
     new_series = check_var(series, "continuous", tmp_path)
     assert new_series.min() > -0.5
@@ -144,6 +154,7 @@ def test_polars_continuous(dtype, tmp_path):
     "dtype", [pl.String, pl.Utf8]
 )
 def test_polars_string(dtype, tmp_path):
+    """Test polars string types."""
     series = pl.Series("series", string.printable, dtype=dtype)
     check_var(series, "string", tmp_path)
 
@@ -151,6 +162,7 @@ def test_polars_string(dtype, tmp_path):
     "dtype", [pl.Categorical]
 )
 def test_polars_categorical(dtype, tmp_path):
+    """Test whether polars categorical data types are parsed correctly."""
     series = pl.Series("series", string.printable, dtype=dtype)
     check_var(series, "categorical", tmp_path)
 
@@ -159,6 +171,7 @@ def test_polars_categorical(dtype, tmp_path):
     [pl.Series, pd.Series],
 )
 def test_float(tmp_path, series_type):
[…]
