+++ tests/test_toml.py
@@ -1,3 +1,4 @@
+"""Tests I/O for toml configuration and GMF files."""
 from filecmp import cmp
 from pathlib import Path
 
@@ -16,7 +17,7 @@ from metasyn.testutils import create_input_toml, create_md_report
 
 
 def test_datafree_create(tmpdir):
-
+    """Test whether a synthetic dataset can be generated without data."""
     temp_toml = tmpdir / "test.toml"
     create_input_toml(temp_toml)
     assert cmp(temp_toml, Path("examples", "config_files", "example_all.toml"))
@@ -33,6 +34,7 @@ def test_datafree_create(tmpdir):
     ]
 )
 def test_toml_save_load(tmpdir, toml_input, data):
+    """Test whether TOML GMF files can be saved/loaded."""
     mf = MetaFrame.fit_dataframe(data, toml_input)
     mf.save(tmpdir/"test.toml")
     new_mf = MetaFrame.load(tmpdir/"test.toml")
@@ -45,11 +47,13 @@ def test_toml_save_load(tmpdir, toml_input, data):
     ]
 )
 def test_md_output(tmpdir, gmf_file):
+    """Test whether the markdown report can be generated from the GMF file."""
     create_md_report(gmf_file, tmpdir/"test.md")
     assert Path(tmpdir/"test.md").is_file()
     assert isinstance(MetaFrame.load(gmf_file), MetaFrame)
 
 def test_toml_err():
+    """Test whether errors are correctly handled."""
     with pytest.raises(FileNotFoundError):
         MetaFrame.from_config("doesnotexist.toml")
 
