### 2024-10-28 286e90b
Make the detection of multinoulli + redacted param
