### 2024-10-29 8923ea2
Initial idea of documentation rewrite
