+++ docs/source/metaframes.rst
@@ -1,0 +1,3 @@
+What are MetaFrames?
+--------------------
+
