+++ docs/source/datafree.rst
@@ -1,0 +1,2 @@
+Synthetic Data without Data
+---------------------------
