### 2024-11-07 44b9857
Update paper with more examples
- added hospital data example
- added drug use data example
