+++ docs/paper/paper.md
@@ -43,36 +43,33 @@ These choices enable the software to generate synthetic data with __privacy and 
 
 # Software features
 
-At its core, `metasyn` has three main functions:
-
-1. __Estimation__: Fit a generative model to a properly formatted tabular dataset, optionally with privacy guarantees.
-2. __(De)serialization__: Create an intermediate representation of the fitted model for auditing, editing, and saving.
-3. __Generation__: Synthesize new datasets based on a fitted model.
+At its core, `metasyn` has three main functions: __estimation__, to fit a model to a properly formatted tabular dataset; __generation__, to synthesize new datasets based on a fitted model; and __(de)serialization__, to create a file from the model for auditing, editing, and saving.
 
 ## Estimation
-The generative model in `metasyn` makes the assumption of marginal independence: each column is considered separately, similar to naïve Bayes classifiers [@hastie2009elements]. Some key advantages of this naïve approach are transparency and explainability, flexibility in handling mixed data types, and computational scalability to high-dimensional datasets. Formally, the generative model for $K$-variate data $\mathbf{x}$ is:
-
-\begin{equation} \label{eq:model}
-p(\mathbf{x}) = \prod_{k = 1}^K p(x_k)
-\end{equation}
-
-Model estimation starts with an appropriately pre-processed data frame, meaning it is tidy [@wickham2014tidy], each column has the correct data type, and missing data are represented by a missing value. Internally, our software uses the `polars` data frame library [@vink2024polars], as it is performant, has consistent data types, and natively supports missing data (i.e., `null` values). An example source table is printed below (NB: categorical data are appropriately encoded as `cat`, not `str`):
+Model estimation starts with an appropriately pre-processed data frame, meaning it is tidy [@wickham2014tidy], each column has the correct data type, and missing data are represented by a missing value. Accordingly, `metasyn` is built on the `polars` data frame library [@vink2024polars]. As an example, the first records of the "hospital" data built into `metasyn` are printed below:
 
 ```
-┌─────┬────────┬─────┬────────┬──────────┐
-│ ID  ┆ fruits ┆ B   ┆ cars   ┆ optional │
-│ --- ┆ ---    ┆ --- ┆ ---    ┆ ---      │
-│ i64 ┆ cat    ┆ i64 ┆ cat    ┆ i64      │
-╞═════╪════════╪═════╪════════╪══════════╡
-│ 1   ┆ banana ┆ 5   ┆ beetle ┆ 28       │
-│ 2   ┆ banana ┆ 4   ┆ audi   ┆ 300      │
-│ 3   ┆ apple  ┆ 3   ┆ beetle ┆ null     │
-│ 4   ┆ apple  ┆ 2   ┆ beetle ┆ 2        │
-│ 5   ┆ banana ┆ 1   ┆ beetle ┆ -30      │
-└─────┴────────┴─────┴────────┴──────────┘
+┌────────────┬───────────────┬───────────────┬──────┬──────┬───────────────┐
+│ patient_id ┆ date_admitted ┆ time_admitted ┆ type ┆ age  ┆ hours_in_room │
+│ ---        ┆ ---           ┆ ---           ┆ ---  ┆ ---  ┆ ---           │
+│ str        ┆ date          ┆ time          ┆ cat  ┆ i64  ┆ f64           │
+╞════════════╪═══════════════╪═══════════════╪══════╪══════╪═══════════════╡
[…]
