### 2024-11-11 81905bd
Improve test documentation (#335)
Also fix some quirkiness in the NA distribution fitting.
