+++ tests/test_dataset.py
@@ -1,3 +1,4 @@
+"""Testing module for fitting dataframes for both pandas and polars."""
 from pathlib import Path
 from random import random
 
@@ -6,7 +7,7 @@ import polars as pl
 import pytest
 from pytest import mark
 
-from metasyn.demo.dataset import _get_demo_class, demo_dataframe, demo_file, _AVAILABLE_DATASETS
+from metasyn.demo.dataset import _AVAILABLE_DATASETS, _get_demo_class, demo_dataframe, demo_file
 from metasyn.metaframe import MetaFrame
 from metasyn.provider import get_distribution_provider
 from metasyn.var import MetaVar
@@ -32,12 +33,14 @@ def _read_csv(fp, dataframe_lib):
         df = pd.read_csv(fp, dtype=dtypes)
         return df.iloc[:100]
     else:
-        df = pl.read_csv(fp, schema_overrides={x: pl.Categorical for x, x_type in dtypes.items() if x_type == "category"})
+        df = pl.read_csv(fp, schema_overrides={x: pl.Categorical for x, x_type in dtypes.items()
+                                               if x_type == "category"})
         return df[:100]
 
 
 @mark.parametrize("dataframe_lib", ["polars", "pandas"])
 def test_dataset(tmp_path, dataframe_lib):
+    """Integration test that loads the CSV, and creates a MetaFrame from that."""
     titanic_fp = Path("tests", "data", "titanic.csv")
     tmp_fp = tmp_path / "tmp.json"
     df = _read_csv(titanic_fp, dataframe_lib)
@@ -82,6 +85,7 @@ def test_dataset(tmp_path, dataframe_lib):
     dataset = MetaFrame.load_json(tmp_fp)
     check_dataset(dataset)
 
+    # Check the description functionality of metasyn.
     dataset.descriptions = {"Embarked": "Some description", "Sex": "Other description"}
     assert dataset.descriptions["Embarked"] == "Some description"
     with pytest.raises(AssertionError):
@@ -106,6 +110,7 @@ def test_dataset(tmp_path, dataframe_lib):
 
 
 def test_distributions(tmp_path):
+    """Create all available distributions and save a metaframe with it."""
     tmp_fp = tmp_path / "tmp.json"
 
     provider = get_distribution_provider()
@@ -120,6 +125,7 @@ def test_distributions(tmp_path):
     "dataset_name", list(_AVAILABLE_DATASETS)
 )
 def test_demo_datasets(tmp_path, dataset_name):
+    """Test all built-in demo datasets and see if they can be synthesized."""
     demo_fp = demo_file(dataset_name)
     demo_df = demo_dataframe(dataset_name)
     demo_class = _get_demo_class(dataset_name)
@@ -141,11 +147,13 @@ def test_demo_datasets(tmp_path, dataset_name):
 
 
 def test_demo_non_exist():
+    """Check that trying to get a demo dataset that doesn't exist raises an error."""
     with pytest.raises(ValueError):
         demo_file("non-existing-dataset")
 
 
 def test_create_dataset(tmpdir):
+    """Test re-creation of the test dataset."""
     test_class = _get_demo_class("test")
     test_class.create(tmpdir/"test.csv")
     df = pl.read_csv(Path(tmpdir/"test.csv"))
