+++ tests/test_var.py
@@ -1,3 +1,7 @@
+"""Module to test the MetaVar object.
+
+Most of the test functions try to infer the correct distribution from a series.
+"""
 import datetime as dt
 import json
 import string
@@ -28,28 +32,33 @@ def _series_drop_nans(series):
     return series.dropna()
 
 
-def _series_element_classname(series):
-    series = _series_drop_nans(series)
+def _series_element_classname(series, all_nan):
+    if not all_nan:
+        series = _series_drop_nans(series)
     if isinstance(series, pl.Series):
         return series[0].__class__.__name__
     return series.iloc[0].__class__.__name__
 
 
-def check_var(series, var_type, temp_path):
+def check_var(series, var_type, temp_path, all_nan=False):
+    """Checker for all variables whether series are inferred correctly."""
     def check_similar(series_a, series_b):
         assert isinstance(series_a, (pd.Series, pl.Series))
         assert isinstance(series_b, (pd.Series, pl.Series))
         assert len(series_a) == len(series_b)
-        base_type_a = _series_element_classname(series_a)
-        base_type_b = _series_element_classname(series_b)
+        base_type_a = _series_element_classname(series_a, all_nan)
+        base_type_b = _series_element_classname(series_b, all_nan)
         if type(series_a) == type(series_b):
             assert base_type_a == base_type_b
-        assert (len(series_a)-len(_series_drop_nans(series_a)) > 0) == (len(series_b) - len(_series_drop_nans(series_b)) > 0)
+        has_nans_a = len(series_a) - len(_series_drop_nans(series_a)) > 0
+        has_nans_b = len(series_b) - len(_series_drop_nans(series_b)) > 0
+        assert has_nans_a == has_nans_b
 
     assert isinstance(series, (pd.Series, pl.Series))
 
     var = MetaVar.fit(series)
     new_series = var.draw_series(len(series))
+    print(new_series)
     check_similar(series, new_series)
     assert var.var_type == var_type
     assert var_type in var.distribution.var_type
@@ -60,7 +69,7 @@ def check_var(series, var_type, temp_path):
     with raises(ValueError):
         var_dict = var.to_dict()
         var_dict.update({"type": "unknown"})
-        MetaVar.from_dict(var_dict)
+        print(MetaVar.from_dict(var_dict))
 
     with raises(ValueError):
         var_dict = var.to_dict()
@@ -99,11 +108,13 @@ def check_var(series, var_type, temp_path):
     ]
 )
 def test_categorical(tmp_path, series):
+    """Test the inference of categorical variables."""
     new_series = check_var(series, "categorical", tmp_path)
     assert set(_series_drop_nans(series)) == set(np.unique(_series_drop_nans(new_series)))
 
 @mark.parametrize("dtype", ["int8", "int16", "int32", "int64", "int"])
 def test_integer(dtype, tmp_path):
+    """Test the inference of integer variables."""
     series = pd.Series([np.random.randint(0, 10) for _ in range(300)], dtype=dtype)
     new_series = check_var(series, "discrete", tmp_path)
     assert new_series.min() >= 0
@@ -112,6 +123,7 @@ def test_integer(dtype, tmp_path):
 
 @mark.parametrize("dtype", ["Int8", "Int16", "Int32", "Int64"])
 def test_nullable_integer(dtype, tmp_path):
+    """Test for pandas nullable integer data types (no applicable to polars)."""
     series = pd.Series([np.random.randint(0, 10) if np.random.rand() > 0.5 else None
[…]
