+++ docs/source/index.rst
@@ -27,57 +27,42 @@ Metasyn Documentation
 =======================
 Welcome to the `metasyn <https://github.com/sodascience/metasyn/>`_ documentation. 
 
-``Metasyn`` is a Python package for generating synthetic tabular data with a focus on privacy. It is designed for owners of sensitive datasets who want to share approximations of their data so that others can perform exploratory analysis and testing without disclosing real values.
+``Metasyn`` is a Python package for generating synthetic tabular data with a focus on privacy.
+It is designed for owners of sensitive datasets who want to share approximations of their data so that
+others can perform exploratory analysis and testing without disclosing real values.
 
-
-Documentation Outline
----------------------
 
 .. toctree::
    :hidden:
    :maxdepth: 2
+   :caption: Getting started
 
-   metasyn_in_detail
+   what_is_metasyn
    installation
    quick_start
-   cli
+
+
+.. toctree::
+   :hidden:
+   :maxdepth: 2
+   :caption: Usage
+
    improve_synth
-   metaframes
    datafree
 
+.. toctree::
+   :hidden:
+   :maxdepth: 2
+   :caption: Advanced usage
+
+   cli
    developer/developer
    api/metasyn
+
+.. toctree::
+   :hidden:
+   :maxdepth: 2
+   :caption: Extras
+
    faq
    about/about
-
-
-This documentation is designed to help you easily navigate and find the information you need. It is organized into the following four sections:
-
-
-:doc:`metasyn_in_detail`
-^^^^^^^^^^^^^^^^^^^^^^^^
-
-The :doc:`/metasyn_in_detail` section provides an overview of metasyn's purpose and functionality.
-
-:doc:`/api/metasyn`
-^^^^^^^^^^^^^^^^^^^^^
-The :doc:`/api/metasyn` is a technical reference for metasyn. Here, each function, class, and module is outlined in detail, giving you a comprehensive understanding of how the package works and how to use its various functionalities.  If, for example, you'd like to discover which parameters can be used for which function, you can find that here.
-
-:doc:`/developer/developer`
-^^^^^^^^^^^^^^^^^^^^^^^^^^^
-The :doc:`/developer/developer` provides resources for those interested in contributing to metasyn's development. This section includes guidance on how to build upon the existing ``metasyn`` codebase, add new modules, functions, or even integrate other packages.
-
-:doc:`/faq`
-^^^^^^^^^^^^^^^^^^^
-The :doc:`/faq` contains commonly asked questions and answers about metasyn. 
-
-:doc:`/about/about`
-^^^^^^^^^^^^^^^^^^^
-The :doc:`About Section </about/about>` provides contact information, and licensing details.
-
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
