+++ docs/source/quick_start.rst
@@ -1,5 +1,5 @@
-Generate synthetic data
-=======================
+Quickstart
+==========
 
 Get started quickly with ``metasyn`` using the following example. In this concise demonstration, you'll learn the basic functionality of ``metasyn`` by generating synthetic data from `titanic <https://raw.githubusercontent.com/pandas-dev/pandas/main/doc/data/titanic.csv>`_ dataset.
 
