+++ docs/source/about.rst
@@ -1,5 +1,49 @@
+About
+=====
+
+Contact Us
+----------
+
+**Metasyn** is a project by the `ODISSEI Social Data Science (SoDa) <https://odissei-data.nl/nl/soda/>`_ team and is currently being maintained by Erik-Jan van Kesteren, Raoul Schram and Samuel Spithorst.
+
+.. image:: /images/logos/soda.png
+   :alt: SODA_logo
+   :width: 200
+   :align: center
+   :target: https://odissei-data.nl/nl/soda/
+
+Do you have questions, suggestions, or remarks on the technical implementation? We welcome your feedback and contributions!
+
+GitHub Repository and Issue Tracker
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+Feel free to check out our `GitHub repository <https://github.com/sodascience/metasyn>`_ for the latest updates and to browse the source code.
+If you encounter any problems or have ideas for improvements, please file an issue on our `GitHub issue tracker <https://github.com/sodascience/metasyn/issues>`_.
+
+.. image:: https://img.shields.io/badge/GitHub-blue?logo=github&link=https%3A%2F%2Fgithub.com%2Fsodascience%2Fmetasyn
+   :alt: GitHub Repository Button
+   :target: https://github.com/sodascience/metasyn
+
+.. image:: https://img.shields.io/badge/GitHub-Issue_Tracker-blue?logo=github&link=https%3A%2F%2Fgithub.com%2Fsodascience%2Fmetasyn%2Fissues   
+   :alt: GitHub Issue Tracker Button
+   :target: https://github.com/sodascience/metasyn/issues
+
+Contributing
+^^^^^^^^^^^^
+
+We highly encourage and appreciate pull requests (PRs) from the community. If you want to discover how you can contribute to metasyn, please refer to our :doc:`/developer/developer`
+
+Maintainers
+^^^^^^^^^^^
+
+Feel free to contact one of the maintainers directly:
+
+* Erik-Jan van Kesteren: `https://github.com/vankesteren <https://github.com/vankesteren>`_
+
+* Raoul Schram: `https://github.com/qubixes <https://github.com/qubixes>`_
+
 License
-========
+-------
 
 Metasyn is released under the MIT License.
 
