+++ docs/source/usage/saving_metaframes.rst
@@ -1,343 +1,0 @@
-.. |break| raw:: html
-
-   <br />
-
-Exporting and importing MetaFrames
-===================================
-
-Metasyn can serialize and **save a MetaFrame** into a GMF file. GMF files are JSON files that follow the :doc:`/developer/GMF` and have been designed to be easy to read and understand. This allows users to audit, understand, modify and share their data generation model with ease. 
-
-.. image:: /images/pipeline_serialization_code.png
-   :alt: MetaFrame Serialization Flow
-   :align: center
-   :width: 400px
-
-Exporting a MetaFrame
-----------------------
-MetaFrames can be serialized and saved to a GMF file by calling the :meth:`metasyn.metaframe.MetaFrame.save` method on a :obj:`MetaFrame<metasyn.metaframe.MetaFrame>`. 
-
-The following code saves a generated :obj:`MetaFrame<metasyn.metaframe.MetaFrame>` object named ``mf`` to a GMF file named ``saved_metaframe``.
-
-.. code-block:: python
-
-   mf.save("saved_metaframe.json")
-
-.. raw:: html
-
-   <details> 
-   <summary> <em><b>An example of a MetaFrame that has been saved to a GMF file: </em></b></summary>
-
-.. code-block:: json
-    
-    {
-        "n_rows": 5,
-        "n_columns": 5,
-        "provenance": {
-            "created by": {
-                "name": "metasyn",
-                "version": "0.8.1"
-            },
-            "creation time": "2024-04-29T12:07:36.646884"
-        },
-        "vars": [
-            {
-                "name": "ID",
-                "type": "discrete",
-                "dtype": "Int64",
-                "prop_missing": 0.0,
-                "distribution": {
-                    "implements": "core.unique_key",
-                    "version": "1.0",
-                    "provenance": "builtin",
-                    "class_name": "UniqueKeyDistribution",
-                    "unique": true,
-                    "parameters": {
-                        "lower": 1,
-                        "consecutive": true
-                    }
-                },
-                "creation_method": {
-                    "created_by": "metasyn",
-                    "unique": true
-                }
-            },
-            {
-                "name": "fruits",
-                "type": "categorical",
-                "dtype": "Categorical(ordering='physical')",
-                "prop_missing": 0.0,
-                "distribution": {
-                    "implements": "core.multinoulli",
-                    "version": "1.0",
-                    "provenance": "builtin",
-                    "class_name": "MultinoulliDistribution",
-                    "unique": false,
-                    "parameters": {
-                        "labels": [
-                            "apple",
-                            "banana"
-                        ],
-                        "probs": [
-                            0.4,
-                            0.6
-                        ]
-                    }
-                },
-                "creation_method": {
-                    "created_by": "metasyn"
-                }
-            },
-            {
-                "name": "B",
-                "type": "discrete",
-                "dtype": "Int64",
-                "prop_missing": 0.0,
-                "distribution": {
-                    "implements": "core.uniform",
-                    "version": "1.0",
-                    "provenance": "builtin",
-                    "class_name": "DiscreteUniformDistribution",
-                    "unique": false,
-                    "parameters": {
-                        "lower": 1,
-                        "upper": 6
-                    }
-                },
-                "creation_method": {
-                    "created_by": "metasyn",
-                    "unique": false
-                }
-            },
-            {
-                "name": "cars",
-                "type": "categorical",
-                "dtype": "Categorical(ordering='physical')",
-                "prop_missing": 0.0,
-                "distribution": {
-                    "implements": "core.multinoulli",
[…]
