+++ docs/source/metaframes.rst
@@ -1,3 +1,0 @@
-What are MetaFrames?
---------------------
-
