+++ docs/source/usage/usage.rst
@@ -1,17 +1,0 @@
-User Guide
-============
-
-Welcome to the ``metasyn`` User Guide. This section features everything you need to know about how to use metasyn. 
-information on installation, quickstart, tutorials and information on the core features of the package.
-
-.. toctree::
-   :maxdepth: 1
-   :caption: Sections:
-
-   generating_metaframes
-   saving_metaframes
-   generating_synthetic_data
-   config_files
-   cli
-   interactive_tutorials
-   extensions

[…]
