+++ docs/source/developer/developer.rst
@@ -4,17 +4,12 @@ Developer Guide
 This guide is mainly directed at developers working on the ``metasyn`` package, but it may be useful for users that want
 more control over the workings of the ``metasyn`` package.
 
-.. warning:: 
-    This section is intended for developers and advanced users. If you are new to ``metasyn``, please refer to the :doc:`/usage/usage` first. 
-
-
 .. toctree::
    :maxdepth: 1
    :caption: Sections:
    
    contributing
    overview
-   GMF
    distributions
    plugins
 

[…]
