### 2024-11-13 633afe5
Run core ci on push (#341)
