### 2024-11-15 2c5b76e
Small fix in hospital dataset description
