### 2024-11-15 1e53df6
Simplified paper description and added relevant examples (#337)
* Update paper with more examples

- added hospital data example
- added drug use data example

* Small fix in hospital dataset description
