### 2024-11-15 672ec64
Small updates to docs
- made index even smaller
- tiny textual changes here and there
- added date/time datatypes to quickstart
- removed binder link (slow!)
