+++ docs/source/what_is_metasyn.rst
@@ -1,9 +1,9 @@
 What is metasyn?
 ================
 
-``Metasyn`` is a python package for generating synthetic data with a focus on maintaining privacy.
+Metasyn is a python package for generating synthetic data with a focus on maintaining privacy.
 It is aimed at owners of sensitive datasets such as public organisations, research groups, and individual researchers who
-want to improve the accessibility, reproducibility and reusability of their data. The goal of ``metasyn`` is to make it easy
+want to improve the accessibility, reproducibility and reusability of their data. The goal of metasyn is to make it easy
 for data owners to share the structure and approximation of the content of their data with others with fewer privacy concerns.
 
 Use cases
@@ -11,11 +11,11 @@ Use cases
 
 Below are a few possible use cases for metasyn:
 
-- A researcher who cannot share their dataset because of privacy concerns, but shares a synthetic twin for reproducibility.
-- A data provider wants to show a preview of the real data using synthetic twins of their datasets.
-- Create synthetic data before you actually have real data so that the analysis scripts can be written and tested.
+- A researcher who cannot share their dataset because of privacy concerns, but shares a synthetic version for reproducibility.
+- A data provider wants to show a preview of their sensitive data using synthetic versions of their datasets.
+- A developer can create synthetic data before they have real data so they can write and test data analysis scripts.
 
-It is specifically **not** designed for creating highly accurate synthetic data, where all relationships between columns are reproduced.
+Metasyn is specifically **not** designed for creating highly accurate synthetic data, where all relationships between columns are reproduced.
 
 Key features
 ------------
@@ -23,10 +23,10 @@ Key features
 - **Fast**: Creating synthetic data takes mere seconds for medium sized (~1000 rows) datasets.
 - **Safe and Understandable Synthetic Data**: As little information as possible is retained from the original dataset, and you can inspect and understand exactly which information is used to create your synthetic data.
 - **Flexible**: You can :doc:`adjust <improve_synth>` the synthetic data columns to your liking, create your own distributions, extension and more.
-- **Data Type Support**: ``Metasyn`` supports generating synthetic data for a variety of common data types including ``categorical``, ``string``, ``integer``, ``float``, ``date``, ``time``, and ``datetime``.
-- **Integration with Faker**: ``Metasyn`` integrates with the `faker <https://github.com/joke2k/faker>`__ package, a Python library for generating fake data such as names and emails. Allowing for more realistic synthetic data.
-- **Structured String Detection**: ``Metasyn`` identifies structured strings within your dataset, which can include formatted text, codes, identifiers, or any string that follows a specific pattern.
-- **Handling Unique Values**: ``Metasyn`` can identify and process variables with unique values or keys in the data, preserving their uniqueness in the synthetic dataset.
+- **Data Type Support**: Metasyn supports generating synthetic data for a variety of common data types including ``categorical``, ``string``, ``integer``, ``float``, ``date``, ``time``, and ``datetime``.
+- **Integration with Faker**: Metasyn integrates with the `faker <https://github.com/joke2k/faker>`__ package, a Python library for generating fake data such as names and emails. Allowing for more realistic synthetic data.
+- **Structured String Detection**: Metasyn identifies structured strings within your dataset, which can include formatted text, codes, identifiers, or any string that follows a specific pattern.
+- **Handling Unique Values**: Metasyn can identify and process variables with unique values or keys in the data, preserving their uniqueness in the synthetic dataset.
 
 
 For more detail on how metasyn works, see our `paper <https://github.com/sodascience/metasyn/blob/main/docs/paper/paper.pdf>`_.
@@ -40,15 +40,15 @@ MetaFrames and GMF files
    :alt: Metasyn Pipeline
    :align: center
 
[…]
