+++ docs/source/installation.rst
@@ -8,16 +8,16 @@ This document will guide you through the process of installing the ``metasyn`` p
 
 
 Step 1: Install Python
----------------------------
+----------------------
 
 Metasyn requires Python 3.8 or higher. If you don't have Python installed or your Python version is lower than 3.8, you will need to install or upgrade Python.
 If you're not familiar with Python or you've never installed it before, refer to the `Python Guide <https://docs.python-guide.org/starting/installation/>`_ for detailed instructions. 
 The Python installation will most likely install `pip`, but if not refer to the `Pip Installation Guide <https://pip.pypa.io/en/stable/installation/>`_ for instructions.
 
 Step 2: Install metasyn
-----------------------------
+-----------------------
 
-Install metasyn from the command prompt with:
+Install metasyn from the command prompt or terminal with:
 
 	.. code-block:: console
 
