+++ docs/source/tutorials.rst
@@ -6,6 +6,5 @@ you can try our interactive tutorial. It provides a step-by-step walkthrough and
 to help you get started.
 
 You can access it here:
-   - Binder: `Open in Binder <https://mybinder.org/v2/gh/sodascience/metasyn/HEAD?labpath=examples%2Fgetting_started.ipynb>`__
    - Google Colab: `Open in Google Colab <https://colab.research.google.com/github/sodascience/metasyn/blob/main/examples/getting_started.ipynb>`__
    - GitHub: `View on GitHub <https://github.com/sodascience/metasyn/blob/main/examples/getting_started.ipynb>`__

[…]
