### 2024-11-15 33e94d9
Update improve and datafree usage sections
