+++ docs/source/improve_synth.rst
@@ -1,104 +1,96 @@
 Improve your synthetic data
 ===========================
 
-When you run metasyn on your dataframe without any directives, metasyn will attempt
-to fit the best distribution to each of your columns. This could be sub-optimal:
-for example, metasyn won't try see whether a column contains names of people for example.
-The column can also be too privacy sensitive to fit with the default methods.
+When you run metasyn on your dataframe, by default it will attempt
+to find the best distribution for each of your columns. This could be sub-optimal:
+for example, metasyn won't know whether a column contains names of people.
+The column can also be too privacy-sensitive to fit with default methods.
 
+Metasyn provides two paths to improving the quality of your synthetic data: by further
+specifying information directly in python, or by providing a configuration file in the
+TOML format. For interactive use, we foresee using python directly, and for programmatic
+use the configuration file is a more appropriate interface (see also our :doc:`cli`).
+
+
+.. tab:: Python
+
+   .. code-block:: python
+
+      from metasyn import MetaFrame, VarSpec
+      from metasyn.distribution import FakerDistribution
+      from metasyncontrib.disclosure import DisclosurePrivacy
+
+      specs = [
+         VarSpec(name="Name", distribution=FakerDistribution(faker_type="name")),
+      ]
+
+      mf = MetaFrame.fit_dataframe(
+         df,
+         privacy=DisclosurePrivacy(),
+         var_specs=specs,
+      )
+
+.. tab:: Configuration file
+
+   .. code-block:: python
+
+      MetaFrame.fit_dataframe(
+         df,
+         var_specs="your_config_file.toml"
+      )
+
+   This refers to a configuration file called ``your_config_file.toml``:
+
+   .. code-block:: toml
+
+      privacy = "disclosure"
+
+      [[var]]
+      name = "Name"
+      description = "Name of the unfortunate passenger of the titanic."
+      distribution = {implements = "core.faker", parameters = {faker_type = "name"}}
+
+   More examples for metasyn configuration files are available on our
+   `GitHub page <https://github.com/sodascience/metasyn/tree/develop/examples/config_files>`_.
 
 .. admonition:: What is the TOML file format?
 
-   The `TOML <https://toml.io/en/>`_ file format can be read with any text editor, and natural to comprehend.
-   You should be able to create your own TOML files from the examples below, but for more details reffer to the TOML 
+   The `TOML <https://toml.io/en/>`_ file format can be read with any text editor, and is natural to comprehend.
+   You should be able to create your own TOML files from the examples below, but for more details refer to the TOML 
    `Documentation <https://toml.io/en/>`_. One important thing to note is that the TOML format is case sensitive.
 
 
-Metasyn provides two ways of improving the quality of your synthetic data: by
-specifying directives directly in the API, or by providing a configuration file in the
-TOML format. Examples for metasyn configuration files are available on our
-`GitHub page <https://github.com/sodascience/metasyn/tree/develop/examples/config_files>`_.
+The remainder of this page serves as a reference for the different options to improve synthetic data quality.
+
 
 General specifications
 ----------------------
 
-There are three general specifications that can be set: ``dist_providers``, ``privacy`` and ``n_rows``.
-Depending on whether you want to use a configuration file, examples are shown below:
-
-.. tab:: Python
-
-   .. code-block:: python
-
[…]
