### 2024-11-15 sodascience/metasyn#326 @vankesteren
@PetrKorab we have extensive automated tests in our [tests](https://github.com/sodascience/metasyn/tree/develop/tests) folder, and we have now additionally documented them with docstrings based on your feedback.

Our tests include:
- Unit tests for different methods on different data types, e.g., [test_continuous.py](https://github.com/sodascience/metasyn/blob/develop/tests/test_continuous.py)
- "integration" tests for all the included datasets in [our datasets folder](https://github.com/sodascience/metasyn/tree/develop/metasyn/demo), see [test_dataset.py](https://github.com/sodascience/metasyn/blob/develop/tests/test_dataset.py)
- tests for the command-line interface, see [test_cli.py](https://github.com/sodascience/metasyn/blob/develop/tests/test_cli.py)

In addition, our continuous integration includes linting, type checking, formatting, documentation building, and running the example jupyter notebook from our [examples](https://github.com/sodascience/metasyn/tree/develop/examples) folder. See [core-ci.yml](https://github.com/sodascience/metasyn/blob/develop/.github/workflows/core-ci.yml)

Data is available to everyone in this way:

```py
from metasyn import demo_dataframe
df = demo_dataframe("hospital")
print(df)
```

```
┌────────────┬───────────────┬───────────────┬──────┬──────┬───────────────┐
│ patient_id ┆ date_admitted ┆ time_admitted ┆ type ┆ age  ┆ hours_in_room │
│ ---        ┆ ---           ┆ ---           ┆ ---  ┆ ---  ┆ ---           │
│ str        ┆ date          ┆ time          ┆ cat  ┆ i64  ┆ f64           │
╞════════════╪═══════════════╪═══════════════╪══════╪══════╪═══════════════╡
│ A5909X0    ┆ 2024-01-01    ┆ 10:30:00      ┆ IVT  ┆ null ┆ 3.633531      │
│ B4025X2    ┆ 2024-01-01    ┆ 11:23:00      ┆ IVT  ┆ 59   ┆ 6.932891      │
│ B6999X2    ┆ 2024-01-01    ┆ 11:58:00      ┆ IVT  ┆ 77   ┆ 1.970654      │
│ B9525X2    ┆ 2024-01-01    ┆ 16:56:00      ┆ MYE  ┆ null ┆ 1.620047      │
│ B0453X1    ┆ 2024-01-02    ┆ 11:50:00      ┆ IVT  ┆ null ┆ 3.344935      │
│ …          ┆ …             ┆ …             ┆ …    ┆ …    ┆ …             │
│ B3830X7    ┆ 2024-01-04    ┆ 16:03:00      ┆ CAT  ┆ 87   ┆ 3.87092       │
│ A7952X3    ┆ 2024-01-04    ┆ 18:00:00      ┆ IVT  ┆ 60   ┆ 4.332882      │
│ A0013X2    ┆ 2024-01-05    ┆ 10:06:00      ┆ CAT  ┆ null ┆ 2.133339      │
│ B2077X8    ┆ 2024-01-05    ┆ 10:43:00      ┆ IVT  ┆ null ┆ 6.026763      │
│ A4700X7    ┆ 2024-01-05    ┆ 14:53:00      ┆ MYE  ┆ 84   ┆ 6.193777      │
└────────────┴───────────────┴───────────────┴──────┴──────┴───────────────┘
```

I hope this answers your comment in a satisfactory way!
