### 2024-11-15 d2b3738
Fix small issues in the documentation
