+++ docs/source/datafree.rst
@@ -1,13 +1,9 @@
 Synthetic data without raw data
 ===============================
 
-<<<<<<< HEAD
-It is also possible to create a GMF file without inputting a dataset, or to add additional fictional columns to those already present in a dataset. 
-=======
 It is also possible to create a GMF file without any input dataset, or to add additional fictive columns to those already present in a dataset. 
->>>>>>> a8a4ebb (Improve data free description with defaults)
 
-To do so, you need to fully specify each column (variable) you want to generate. You will also need to set the data_free parameter to true,
+To do so, you need to fully specify each column (variable) you want to generate. You will also need to set the ``data_free`` parameter to true,
 to indicate that the variable will be generated from scratch, instead of being based on existing data.
 Finally, you will need to set the number of rows to generate.
 
@@ -49,7 +45,7 @@ Below is an example on how to set defaults:
    string = {implements = "core.faker", parameters = {faker_type = "name", locale = "en_US"}}
 
 With this block, you won't have to set the ``data_free`` parameter, and the default
-proportion of missing values is set to 0.1, for discrete columns, the distribution
+proportion of missing values is set to 0.1. For discrete columns, the distribution
 will be set to a uniform distribution between 1 and 30, etc.
 
 With the defaults set as above, you only need to specify the ``name`` and ``var_type`` of
