### 2024-11-15 04d3901
Documentation update 1.1 (#340)
* General readthedocs documentation rewrite to make things easier to find.
* Fix bug in assigning privacy
