+++ docs/source/metasyn_in_detail.rst
@@ -1,296 +1,0 @@
-How does metasyn work?
-======================
-
-``Metasyn`` is a python package for generating synthetic data with a focus on maintaining privacy. It is aimed at owners of sensitive datasets such as public organisations, research groups, and individual researchers who want to improve the accessibility, reproducibility and reusability of their data. The goal of ``metasyn`` is to make it easy for data owners to share the structure and approximation of the content of their data with others with fewer privacy concerns.
-
-With this goal in mind, ``metasyn`` restricts itself to the `'synthetically-augmented plausible' <https://www.ons.gov.uk/methodology/methodologicalpublications/generalmethodology/onsworkingpaperseries/onsmethodologyworkingpaperseriesnumber16syntheticdatapilot>`__ category of synthetic data, as categorized by the Office for National Statistics (ONS).
-
-.. admonition:: ONS criteria for a Synthetically-augmented plausible dataset:
-   
-  * Preserve the format and record-level plausibility as detailed previously and replicate marginal (univariate) distributions where possible.
-  * Constructed based on the real dataset, values are generated based on observed distributions (with added fuzziness and smoothing) but no attempt made to preserve relationships.
-  * Missing value codes and their frequency is to be preserved.
-  * Disclosure control evaluation is necessary case by case, special care to be taken with names and so on.
-  * To be used for extended code testing, minimal analytical value, non-negligible disclosure risk.
-
-
-The ``metasyn`` package also incorporates a **plug-in** system, which enables implementations to generate synthetic data with
-stricter and formal privacy and disclosure guarantees within the same framework. Moreover, our system provides an **auditable and editable intermediate representation** in the form of a human- and machine-readable ``.json`` metadata file from which new data can be synthesized.
-
-Through our focus on privacy and transparency, ``metasyn`` explicitly avoids generating synthetic data with high analytical validity. The data generated by our system is realistic in terms of data structure and plausible in terms of values for each variable, but any multivariate relations or conditional patterns are excluded. This has implications for how this synthetic data can be used: not for statistical analysis and inference, but rather for initial exploration, analysis script development, and communication outside the data owner’s institution. In the intended use case, external researchers can make use of the synthetic data to assess the feasibility of their intended research before making the (often time-consuming) step of requesting access to the sensitive source data for the final analysis.
-
-The ``metasyn`` Pipeline
-------------------------
-The three key stages of the ``metasyn`` pipeline include the **estimation** of the MetaFrame from the original data, the **serialization** of the MetaFrame into an auditable and editable intermediate representation, and the **generation** of the synthetic data from the model represented by the MetaFrame or its serialized representation. This section provides a walkthrough of these steps.
-
-.. image:: /images/pipeline_basic.png
-   :alt: Metasyn Pipeline
-   :align: center
-
-.. admonition:: An example use case
-
-  The following illustrates a simple example of ``metasyn`` being used, and where the steps in the pipeline (estimation, serialization/deserialization and generation) occur. 
-  
-  A public health researcher, aiming to conduct research on a sensitive dataset could:
-
-    #. Conduct statistical research on a sensitive dataset of medical records.
-    #. Fit a MetaFrame to the sensitive dataset. (Estimation)
-    #. Export the MetaFrame to a JSON file following the GMF standard. (Serialization)
-    #. Check the GMF file to ensure no private information is present anymore.
-    #. Use the MetaFrame to generate a synthetic dataset. (Generation)
[…]
