+++ docs/source/about/about.rst
@@ -1,11 +1,0 @@
-About
-=====
-
-Welcome to the ``metasyn`` About section. This section is meant to provide more information on the background and context of metasyn. 
-
-.. toctree::
-   :maxdepth: 1
-   :caption: Sections:
-   
-   contact
-   license
