### 2024-11-19 066fca5
Remove stray print statement and add comment.
