### 2024-11-20 #7099 @PetrKorab
@vankesteren Thanks for adding the tests, domain datasets and updating the paper!
