### 2024-11-25 57f5bc1
Bump version to 1.1.0
