+++ docs/source/conf.py
@@ -25,7 +25,7 @@ author = "SODA Team"
 
 
 # The full version, including alpha/beta/rc tags
-release = '1.0.3'
+release = '1.1.0'
 
 # -- General configuration ---------------------------------------------------
 
