### 2024-11-25 f79d562
Update scheduled-ci.yml
