+++ .github/workflows/scheduled-ci.yml
@@ -68,4 +68,4 @@ jobs:
       run: |
         pip install git+https://github.com/sodascience/metasyn-disclosure-control
         pip install .
-        metasyn create-meta metasyn/demo/demo_titanic.csv --config examples/example_config.toml
+        metasyn create-meta metasyn/demo/demo_titanic.csv --config examples/config_files/example_config.toml
