### 2024-11-25 2d90439
Improve error messaging for variable fitting
