### 2024-11-25 2991370
Fix erroneous mypy error
