### 2024-11-27 351120d
update dockerfile for dependency management
