### 2024-11-27 253aab0
Fix numpy dependency, move to stable, add 3.13 ci
