### 2024-11-28 25986a2
Fix numpy dependency, move to stable, add 3.13 ci (#348)
