+++ .github/workflows/scheduled-ci.yml
@@ -12,7 +12,7 @@ jobs:
       fail-fast: false
       matrix:
         branch: ["main", "develop"]
-        python-version: ["3.9", "3.10", "3.11", "3.12"]
+        python-version: ["3.9", "3.10", "3.11", "3.12", "3.13"]
     runs-on: "ubuntu-latest"
     steps:
       - uses: actions/checkout@v4
