### 2024-12-02 6d28014
Make the vartype determination easier
