### 2024-12-03 #7099 @misken
In the left sidebar for the updated paper, the links to the Review and
Repository aren't pointing to the correct places (or I'm missing something).

On Fri, Nov 15, 2024 at 6:28 AM Erik-Jan van Kesteren <
***@***.***> wrote:

> @PetrKorab <https://github.com/PetrKorab> I have now left a comment on
> your main issue on this, here: sodascience/metasyn#326 (comment)
> <https://github.com/sodascience/metasyn/issues/326#issuecomment-2478601871>
>
> We have also updated the paper, making it more legible and with better
> examples based on reviewer feedback, while keeping it as short as possible.
> The new paper is available here:
> https://github.com/sodascience/metasyn/blob/develop/docs/paper/paper.pdf
>
> We are currently in the process of updating the documentation once more
> and then we will release Metasyn version 1.1! I will ping this thread when
> we do our release, that would be a good moment to have another "round" of
> reviews (where hopefully we will check all the boxes 😉)
>
