### 2024-12-03 89c0eff
Split varspec parameter into config
