### 2024-12-03 ac2bcb6
Update documentation and mypy/ruff
