### 2024-12-03 #7099 @misken
My checklist is complete. I've used the software locally but did not do a dev install to run the tests - I'm an Anaconda user and ran into some issues. I'm fine with recommending publication at this point.
