### 2024-12-18 874150e
Allow the synthesized output to be seeded
