+++ metasyn/var.py
@@ -232,7 +232,7 @@ class MetaVar:
             return None
         return self.distribution.draw()
 
-    def draw_series(self, n: int) -> pl.Series:
+    def draw_series(self, n: int, seed: Optional[int]) -> pl.Series:
         """Draw a new synthetic series from the metadata.
 
         Parameters
@@ -245,6 +245,8 @@ class MetaVar:
         polars.Series:
             Polars series with the synthetic data.
         """
+        if seed is not None:
+            np.random.seed(seed)
         self.distribution.draw_reset()
         value_list = [self.draw() for _ in range(n)]
         pl_type = self.dtype.split("(")[0]
