### 2024-12-19 bce1e04
Split varspec param (#351)
* Split varspec parameter into config

* Update documentation and mypy/ruff
