### 2024-12-19 01a1280
Add test to CI
