+++ .github/workflows/core-ci.yml
@@ -65,10 +65,11 @@ jobs:
         pip install .
         metasyn create-meta metasyn/demo/demo_titanic.csv --config examples/config_files/example_config.toml
     - name: Test GMF file
+      if: ${{ matrix.os == 'ubuntu-latest' }}
       run: |
         metasyn schema -o current_schema.json
         MD5_DATA=(`md5sum current_schema.json`)
-        [[ ${MD5_DATA[0]} == "ca057b145b035ff1f2b4032d414d1be9" ]]
+        [[ ${MD5_DATA[0]} == "c2a69330cec7a147ab775f5f0e037d8b" ]]
 
 
 
