### 2024-12-19 0eadc64
Create a new option to output the schema to a file (#352)
* Also add the schema to the CI, so we know when the schema has changed.
