### 2024-12-20 e2fa411
Add/fix more tests
