### 2024-12-20 3740e8b
Add docstring for new function
