### 2024-12-20 624759a
Add to CLI and add documentation
