### 2024-12-20 f82e66d
Allow the synthesized output to be seeded (#353)
