+++ metasyn/var.py
@@ -10,6 +10,7 @@ import polars as pl
 from metasyn.distribution.base import BaseDistribution
 from metasyn.privacy import BasePrivacy, BasicPrivacy
 from metasyn.provider import BaseDistributionProvider, DistributionProviderList
+from metasyn.util import set_global_seeds
 from metasyn.varspec import DistributionSpec
 
 
@@ -232,7 +233,7 @@ class MetaVar:
             return None
         return self.distribution.draw()
 
-    def draw_series(self, n: int) -> pl.Series:
+    def draw_series(self, n: int, seed: Optional[int]) -> pl.Series:
         """Draw a new synthetic series from the metadata.
 
         Parameters
@@ -245,6 +246,9 @@ class MetaVar:
         polars.Series:
             Polars series with the synthetic data.
         """
+        if seed is not None:
+            set_global_seeds(seed)
+
         self.distribution.draw_reset()
         value_list = [self.draw() for _ in range(n)]
         pl_type = self.dtype.split("(")[0]
