+++ tests/test_var.py
@@ -48,16 +48,22 @@ def check_var(series, var_type, temp_path, all_nan=False):
         assert len(series_a) == len(series_b)
         base_type_a = _series_element_classname(series_a, all_nan)
         base_type_b = _series_element_classname(series_b, all_nan)
-        if type(series_a) == type(series_b):
+        if type(series_a) is type(series_b):
             assert base_type_a == base_type_b
         has_nans_a = len(series_a) - len(_series_drop_nans(series_a)) > 0
         has_nans_b = len(series_b) - len(_series_drop_nans(series_b)) > 0
         assert has_nans_a == has_nans_b
 
+    def check_random_draw(var, n_series):
+        series_1 = var.draw_series(n_series, 1234)
+        series_2 = var.draw_series(n_series, 1234)
+        assert all(_series_drop_nans(series_1) == _series_drop_nans(series_2))
+
     assert isinstance(series, (pd.Series, pl.Series))
 
     var = MetaVar.fit(series)
-    new_series = var.draw_series(len(series))
+    new_series = var.draw_series(len(series), 5123)
+    check_random_draw(var, len(series))
     print(new_series)
     check_similar(series, new_series)
     assert var.var_type == var_type
@@ -76,10 +82,10 @@ def check_var(series, var_type, temp_path, all_nan=False):
         var_dict["distribution"].update({"implements": "unknown"})
         MetaVar.from_dict(var_dict)
 
-    newer_series = new_var.draw_series(len(series))
+    newer_series = new_var.draw_series(len(series), 6789)
     check_similar(newer_series, series)
 
-    assert type(new_var) == type(var)
+    assert type(new_var) is type(var)
     assert new_var.dtype == var.dtype
     assert var_type == new_var.var_type
 
@@ -90,9 +96,9 @@ def check_var(series, var_type, temp_path, all_nan=False):
 
     with open(tmp_fp, "r") as f:
         new_var = MetaVar.from_dict(json.load(f))
-    check_similar(series, new_var.draw_series(len(series)))
+    check_similar(series, new_var.draw_series(len(series), 8234))
 
-    assert type(new_var) == type(var)
+    assert type(new_var) is type(var)
     assert new_var.dtype == var.dtype
     assert new_var.var_type == var_type
     assert new_var.creation_method["created_by"] == "metasyn"
@@ -220,7 +226,7 @@ def test_bool(tmp_path, series_type):
     series = series_type(np.random.choice([True, False], size=100))
     check_var(series, "categorical", tmp_path)
     var = MetaVar.fit(series)
-    new_series = var.draw_series(10)
+    new_series = var.draw_series(10, 1234)
     assert new_series.dtype == pl.Boolean
 
 
