+++ metasyn/util.py
@@ -1,6 +1,11 @@
 """Utility module for metasyn."""
 from __future__ import annotations
 
+import random
+
+import faker
+import numpy as np
+
 try:
     import tomllib
 except ImportError:
@@ -27,3 +32,15 @@ def get_registry() -> dict:
     with open(registry_fp, "rb") as handle:
         registry = tomllib.load(handle)
     return registry
+
+def set_global_seeds(seed: int):
+    """Set the global seeds for all random number generators.
+
+    Parameters
+    ----------
+    seed
+        The seed to use for the random number generators.
+    """
+    np.random.seed(seed)
+    random.seed(seed)
+    faker.Faker.seed(seed)
