+++ tests/test_string.py
@@ -4,17 +4,22 @@ import polars as pl
 from pytest import mark
 
 from metasyn.distribution.string import FakerDistribution, FreeTextDistribution
+from metasyn.var import MetaVar
 
 
 @mark.parametrize("series_type", [pd.Series, pl.Series])
 def test_faker(series_type):
     """Test the faker distribution."""
-    var = FakerDistribution.fit(series_type([1, 2, 3]))
-    assert isinstance(var.to_dict(), dict)
-    assert isinstance(var.draw(), str)
-    assert 'faker' in str(var)
-    assert var.locale == "en_US"
-    assert var.faker_type == "city"
+    dist = FakerDistribution.fit(series_type([1, 2, 3]))
+    assert isinstance(dist.to_dict(), dict)
+    assert isinstance(dist.draw(), str)
+    assert 'faker' in str(dist)
+    assert dist.locale == "en_US"
+    assert dist.faker_type == "city"
+    var = MetaVar("some_city", "string", dist, prop_missing=0.0)
+    series_1 = var.draw_series(100, seed=1234)
+    series_2 = var.draw_series(100, seed=1234)
+    assert all(series_1 == series_2)
 
 
 @mark.parametrize(
@@ -35,4 +40,7 @@ def test_free_text(series, lang, avg_sentences, avg_words):
     assert dist.locale == lang
     assert dist.avg_sentences == avg_sentences
     assert dist.avg_words == avg_words
-    dist.draw()
+    var = MetaVar("some_var", "string", dist, prop_missing=0.0)
+    series_1 = var.draw_series(100, seed=1234)
+    series_2 = var.draw_series(100, seed=1234)
+    assert all(series_1 == series_2)
