### 2024-12-24 #7099 @crvernon
👋  @vankesteren - we are almost there! Next is just setting up the archive for your new release.

So here is what we have left to do:

- [ ]  I am assuming that your current version is the final (v1.1.0) for which you have also created a Zenodo archive for.  Please edit the metadata of the Zenodo archive to ensure that the title matches the title of the paper exactly.

In the paper:

- [ ] LINE 144:  The "b" in "bayesian" should be capitalized.  You can do this by placing characters you wish to maintain formatting of in your bib file in curly brackets.
- [ ] LINE 148:  "r" should be capitalized
- [ ] LINE 156:  "r" should be capitalized

Thats it.  Let me know once these changes have been made and I'll move to accept this for publication.  Thanks!
