### 2025-01-03 baf029d
Final paper edits for JOSS
based on https://github.com/openjournals/joss-reviews/issues/7099#issuecomment-2561190547
