### 2025-01-03 c956384
add small ignore type for mypy
