### 2025-01-03 #7099 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
